### 2024-12-15 #7591 @warrickball
Hi @xiaziyna — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, I notice that your co-authors aren't listed as having contributed to the code. Could you briefly describe here what their contributions to the project have been? We don't object to authors who are not code contributors but do have [requirements on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship), which I'll quote here for convenience:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.
