### 2024-12-16 #7591 @xiaziyna
Hi @warrickball, co-authors contributed to developing research ideas implemented here (optical methods/model) and advised on the code use cases and computational methods.
