+++ examples/exovista_sim.py
@@ -17,6 +17,7 @@ from matplotlib.colors import LogNorm
 import os
 
 fname = '1215-HIP_16537-TYC_-mv_3.72-L_0.35-d_3.20-Teff_5048.07.fits'
+
 script_dir = os.path.dirname(os.path.abspath(__file__))
 data_dir = os.path.abspath(os.path.join(script_dir, '..', 'pystarshade', 'data'))
 file_path = os.path.join(data_dir, 'scenes', fname)
@@ -27,13 +28,15 @@ t = 0 # timestep
 scene = Scene(file_path)
 bg_scene = scene.getDiskImage()
 star_xy = scene.getXYstar()
-bg_scene[:,int(star_xy[0]), int(star_xy[1])] = scene.getStarSpec(time = t)
+#bg_scene = np.zeros(np.shape(bg_scene))
+bg_scene[:,int(star_xy[0]), int(star_xy[1])] += scene.getStarSpec(time = t)
 
 planet_xy = scene.getXYplanets(time = t).astype(int)
 planet_spectra = scene.getPlanetSpec(time = t)
 star_spectra = scene.getStarSpec(time = t)
 
 ds_mas = scene.getPixScale()
+print (ds_mas)
 stel_ang_diam = scene.getAngDiam()
 ds = dist_xo_ss * ds_mas * mas_to_rad
 ds_au = ds / au_to_meter
@@ -52,6 +55,7 @@ extended_field[:, half_N - half_in_N : half_N + half_in_N + bit_arr,
        half_N - half_in_N : half_N + half_in_N + bit_arr] = bg_scene
 
 n_planet = 0
+
 for i in range(1, scene.nplanets):
     if np.all(planet_xy[i] - N_dust_half + half_N  <= N_s-1) and np.all(planet_xy[i] - N_dust_half + half_N >= 0):
         extended_field[:, planet_xy[i, 1]- N_dust_half + half_N , planet_xy[i, 0]- N_dust_half + half_N] += planet_spectra[i]
@@ -59,14 +63,27 @@ for i in range(1, scene.nplanets):
 
 w_i = 154 # choose wavelength
 source_field = extended_field[w_i]
+
+plt.figure()
+plt.title('ExoVista input scene', fontsize=14)
+plt.text(-940, 900, r'$\lambda = $500 nm', color='white', fontsize=12)
+plt.imshow(np.abs(source_field),  norm=LogNorm(), cmap='jet')
+plt.xlabel('x [mas]', fontsize=14)
+plt.ylabel('y [mas]', fontsize=14)
+cbar = plt.colorbar()
+cbar.set_label('Contrast', fontsize=14)
+plt.tight_layout()
+plt.show()
+
+
 wl = wl_range[w_i]
 
-print ('Distance from exo-scene to starshade [pc]: ', dist_xo_ss)
+print ('Distance from exo-scene to starshade [pc]: ', dist_xo_ss/pc_to_meter)
 print ('Number of planets included in scene: ', n_planet)
 print ('Star brightness: ', star_spectra[w_i])
 print ('Planet brightness: ', planet_spectra[1:,w_i])
 print ('Plan/Star brightness: ', (planet_spectra[1:,w_i]/(star_spectra[w_i])))
-print ('Semi major axis[au]: ', ds_au*np.hypot(planet_xy[1:, 0] - N_dust_half, planet_xy[1:, 1] - N_dust_half))
+print ('Semi major axis[au]: ', np.sqrt(2)*ds_au*np.hypot(planet_xy[1:, 0] - N_dust_half, planet_xy[1:, 1] - N_dust_half))
 print ('wavelength [m]', wl)
 print ('ZODIs: ', hdul[2].header['NZODIS-0'])
[…]
