+++ examples/earth_sim.py
@@ -4,7 +4,7 @@ Uses 'Scene.py' available in ExoVista directory on Github:
 github.com/alexrhowe/ExoVista
 
 Scenes generated with ExoVista are used here
-Each pixel is 1 mas spacing, with 1001x1001 scene
+Each pixel is 2 mas spacing, with 1001x1001 scene
 """
 from astropy.io import fits
 from . import download_exovista
@@ -17,10 +17,31 @@ from matplotlib.colors import LogNorm
 from matplotlib import rc
 rc('text', usetex=True)
 rc('font',**{'family':'serif','serif':['Times']})
-
 import os
 
-fname = '1215-HIP_16537-TYC_-mv_3.72-L_0.35-d_3.20-Teff_5048.07.fits'
+def flux_to_mag(flux):
+    # 2 mas pixels
+    return - 2.5*np.log10((500**2)*flux /  3640	)
+
+def ring_mean(image, center, radius, thickness=1.0):
+    cy, cx = center
+    ny, nx = image.shape
+    
+    x = np.arange(nx)
+    y = np.arange(ny)
+    xx, yy = np.meshgrid(x, y)
+    
+    rr = np.sqrt((xx - cx)**2 + (yy - cy)**2)
+    
+    inner_radius = radius - thickness / 2
+    outer_radius = radius + thickness / 2
+    ring_mask = (rr >= inner_radius) & (rr < outer_radius)
+    
+    return image[ring_mask].mean(), ring_mask
+
+
+#fname = '999-HIP_-TYC_SUN-mv_4.83-L_1.00-d_10.00-Teff_5778.60.fits' #solar system at 60 deg inclinati
+fname = '999-HIP_-TYC_SUN-mv_4.83-L_1.00-d_10.00-Teff_5778.00.fits'
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 data_dir = os.path.abspath(os.path.join(script_dir, '..', 'pystarshade', 'data'))
@@ -28,11 +49,14 @@ file_path = os.path.join(data_dir, 'scenes', fname)
 hdul = fits.open(file_path)
 dist_xo_ss = hdul[4].header['dist'] * pc_to_meter
 
+inclination = hdul[2].header['I-0']
+print (hdul[2].header['I-0'], hdul[2].header['NZODIS-0'], hdul[2].header['R-0'])
+local_zodi_mag_v_arcsec2 = 22.5
+
 t = 0 # timestep
 scene = Scene(file_path)
 bg_scene = scene.getDiskImage()
 star_xy = scene.getXYstar()
-#bg_scene = np.zeros(np.shape(bg_scene))
 bg_scene[:,int(star_xy[0]), int(star_xy[1])] += scene.getStarSpec(time = t)
 
 planet_xy = scene.getXYplanets(time = t).astype(int)
@@ -40,16 +64,18 @@ planet_spectra = scene.getPlanetSpec(time = t)
 star_spectra = scene.getStarSpec(time = t)
 
 ds_mas = scene.getPixScale()
+
 stel_ang_diam = scene.getAngDiam()
 ds = dist_xo_ss * ds_mas * mas_to_rad
 ds_au = ds / au_to_meter
+
 wl_range = scene.lambdas * 1e-6 
 
 N_wl = len(wl_range)
 N_dust = np.shape(bg_scene)[1]
 N_dust_half = N_dust//2
 
-N_s = 1001
+N_s = N_dust
 extended_field = np.zeros((N_wl, N_s, N_s), dtype=np.complex128) 
 half_N = N_s // 2
 half_in_N = N_dust // 2
@@ -57,20 +83,34 @@ bit_arr = N_dust % 2
[…]
