### 2025-03-11 #7591 @xiaziyna
Hi @ivastar, I look forward to working with you. Does the reviewer have to be listed on the JOSS database of reviewers? If so, I can suggest the following: Iva Laginja @ivalaginja, Jaren Nicholas Ashcraft
 @Jashcraf, Ewan Douglas @douglase.
