### 2025-03-18 #7917 @editorialbot
[REVIEW]: PyStarshade: simulating high-contrast imaging of exoplanets with starshades
**Submitting author:** @xiaziyna (<a href="http://orcid.org/0000-0001-5475-1975">Jamila Taaki</a>)
**Repository:** https://github.com/xiaziyna/PyStarshade
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.8
