### 2025-03-18 #7917 @ivastar
@Jashcraf, @ivalaginja thank you for your willingness to review this submission! Instructions are above. I believe you both have experience reviewing for JOSS before so you know how this works. But please don't hesitate to reach out to me if you have any questions or concerns. We are looking for a first round of reviews if ~3 weeks, so let's say April 10th for this submission. Let me know if you would need more time.
