### 2025-03-24 xiaziyna/PyStarshade#4 @Jashcraf
Examples require usage of pickled objects
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

Hello, thanks for having me as a reviewer - super cool to see an optical propagation library tailored to modeling starshades!

I was running the examples to verify functionality, and encountered the following in `visualize_throughput.py`

```bash
PyStarshade % python examples/visualize_throughput.py
Traceback (most recent call last):
  File "/Users/kaladin/joss_reviews/PyStarshade/examples/visualize_throughput.py", line 22, in <module>
    data = np.load(file_path_on)
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/_npyio_impl.py", line 486, in load
    raise ValueError(
ValueError: This file contains pickled (object) data. If you trust the file you can load it unsafely using the `allow_pickle=` keyword argument or `pickle.load()`.
```

Unless I'm mistaken, it looks like the HWO pupil models are written as binaries using the `.npz` extension, and that's what is being "un-pickled" here. The error above to prevent users from executing pickled data is because [pickling is unsafe](https://docs.python.org/3/library/pickle.html).

I would suggest switching the data required to use `PyStarshade` to another format so that users don't need to worry about pickles (what a funny world we live in). I don't have a particular preference on what `PyStarshade` uses, but here are some options:
- .txt or .csv by loading / saving using `np.savetxt` / `np.loadtxt`, `pandas` dataframes, or something similar
- .fits files, which are [a commonly-used file format in astronomy / astronomical instrumentation](https://docs.astropy.org/en/latest/io/fits/index.html) if you didn't already know about them
