### 2025-03-26 xiaziyna/PyStarshade#4 @xiaziyna
Hi Jaren! Thanks for agreeing to review PyStarshade :)

From what I can tell, if an .npz file only contains Numpy arrays, they do not get pickled. If saving dicts or objects, they get pickled.

I have checked all files in PyStarshade (as well as those in this example) load with the 'allow_pickle = False' setting in np.load, and also made sure they only contain numpy arrays. I have also tried a fresh install.

Can you tell me any other info around this error (did you create a new pupil file/did you git clone and use LFS to download the files)?
Thanks - I'm sorry about this!
