### 2025-04-07 xiaziyna/PyStarshade#4 @Jashcraf
Sorry for dropping off of this - had a bit of a work sprint at an observatory and I dropped every other ball I was "juggling"

Ah I think my error was not installing the files via LFS - so `np.load()` was looking for a file that didn't exist and threw the pickle error. I'm currently downloading the files now and will re-run the examples to verify functionality. Will close issue once I've run all the examples.
