### 2025-04-07 xiaziyna/PyStarshade#5 @Jashcraf
Section for pre-computed data installation instructions on readthedocs
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

The instructions for installing the pre-computed data to run the examples are in the README.md but not present on the readthedocs. Could this information be added on the landing page under "Installation"?

If possible, I'd also request a warning in this section about the file size when running `git lfs pull`. It's taking me quite some time to download the data - looks like the first file out of 33 is several gigabytes in size?
