### 2025-04-07 xiaziyna/PyStarshade#6 @Jashcraf
Request for tutorial on integration with HCIPy
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

In the software paper, the following claim is made

```markdown
‘PyStarshade’ optionally interfaces with HCIPy (Por et al., 2018) to generate telescope apertures and the resulting imaging in a streamlined fashion.
```

This is an cool feature that is worth highlighting in either the examples or the readthedocs.  Could a tutorial be added showing how to use the HCIPy apertures in PyStarshade?

Also, I see that apodization/pupil.py calls the various HCIPy aperture generators: https://github.com/xiaziyna/PyStarshade/blob/0bfc46c23c87c2542c42ba2ad7d617208c2131b2/pystarshade/apodization/pupil.py#L12-L28

But I'm not clear on what "the resulting imaging in a streamlined fashion" refers to. Can PyStarshade export simulated images to HCIPy's `Wavefront` or `Field` object?
