### 2025-04-07 xiaziyna/PyStarshade#4 @Jashcraf
No longer getting the pickle issue! Closing
