### 2025-04-07 xiaziyna/PyStarshade#7 @Jashcraf
Error when running examples/visualize_throughput.py
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

To install git lfs and get the necessary data, I did the following in a fresh conda environment
```bash
brew install git-lfs
pip install pystarshade
git clone https://github.com/xiaziyna/PyStarshade.git PyStarshade
cd PyStarshade
git lfs pull
```

And I see the `data` directory has appeared in `pystardshade`.

Running `visualize_throughput.py` I get the following error on line 59:

```bash
Traceback (most recent call last):
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/format.py", line 677, in _read_array_header
    dtype = descr_to_dtype(d['descr'])
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/format.py", line 326, in descr_to_dtype
    return numpy.dtype(descr)
TypeError: data type '<f16' not understood

Traceback (most recent call last):
  File "/Users/kaladin/joss_reviews/PyStarshade/examples/visualize_throughput.py", line 59, in <module>
    wl = data_throughput_on['wl']
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/_npyio_impl.py", line 254, in __getitem__
    return format.read_array(bytes,
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/format.py", line 804, in read_array
    shape, fortran_order, dtype = _read_array_header(
  File "/opt/anaconda3/envs/joss/lib/python3.10/site-packages/numpy/lib/format.py", line 680, in _read_array_header
    raise ValueError(msg.format(d['descr'])) from e
ValueError: descr is not a valid dtype descriptor: '<f16'
(joss) kaladin@jarens-mbp PyStarshade %
```

Looks like I can successfully load the data_throughput_on, and see that it has the right `wl` key, but cannot access the key because of the above error.
```python
ipdb> data_throughput_on
NpzFile '/Users/kaladin/joss_reviews/PyStarshade/pystarshade/data/psf/hwo_throughput_hwopupil_onaxis_001m.npz' with keys: core_throughput, total_throughput, grid_points, wl, d_pix_mas
```

```python
ipdb> data_throughput_on['wl']
*** ValueError: descr is not a valid dtype descriptor: '<f16'
```

I'm assuming this is a `numpy.load` error? I'm currently running `numpy==2.2.4`. Do you see the same error when running this example?
