### 2025-04-09 xiaziyna/PyStarshade#7 @xiaziyna
Thanks for catching this.

This error was caused because I was using the dtype float128 for 'wl' and it is only supported on x86_64/may have problems with Mac OS. Using float128 is most definitely overkill - I have converted the 'wl' in affected files (just 3 files) to float64 and changed 'propagator.py' to use float64.

The following should work to update the repo:

```bash
git pull
git lfs pull
```
