### 2025-04-09 xiaziyna/PyStarshade#5 @xiaziyna
Hi!
I have updated the docs with additional install info and a warning about the data-size.
