### 2025-04-09 xiaziyna/PyStarshade#7 @Jashcraf
Ah yeah I should have specified that I'm running on ARM - excellent catch! I'll give it a run later today.
