### 2025-04-09 xiaziyna/PyStarshade#7 @Jashcraf
Successful generation of a very cool plot! @xiaziyna one thing I noticed is that "Total Throughput" goes above 1. Is this expected behavior? Or is the "Total throughput" perhaps being normalized erroneously?
