### 2025-04-09 xiaziyna/PyStarshade#7 @xiaziyna
Hi @Jashcraf - this is actually expected behaviour - starshades seem to amplify intensity at ~2*IWA. See Fig 3. in 'A new approach for the numerical calculation of diffraction patterns using starshades, Aime et al. 2024' for similar results.
Here I've normalized by the case when no starshade is present - let me know if you think this conflicts with the conventional definition!
