### 2025-04-10 xiaziyna/PyStarshade#8 @ivalaginja
Installation cleanup
This is part of https://github.com/openjournals/joss-reviews/issues/7917.

Going through the installation instructions, the first thing I asked myself when seeing the optional LFS pull after a git clone was: What can I run with this package if I *don't* do that? Digging into this, I found a couple of issues that would need to be addressed to make the installation process more robust.

## Some comments:
- When cloning the repo for the files but installing the package from PyPI, I now have two separate instances of your code on my machine that are not in sync. Even if I were to go into the local repo clone and change some code there, it would still import it from the pip installation from PyPI, which could lead to confusion.
- Since the assumption is that the package installation happens from PyPI, the default output path across the repository saves output data into the installed module, which is not desirable. Data outputs and code should be kept separate. It looks like @jashcraf ran into the same issue: https://github.com/xiaziyna/PyStarshade/issues/7.
- Your `setup.py` assumes Python versions that have reached or are about to reache end-of-life (EOL): https://devguide.python.org/versions/. This leads to some dependency confusion when running in newer Python versions.
- It's unclear which examples require the provided data and which not.
- it is unclear in which cases the optional dependencies `astropy` and `hcipy` are actually necessary.

## Requested revisions:

### PyPI installation vs. cloning the repo

1. Separate your two installation pathways very clearly: one that installs from PyPI and allows only running your package without the provided data, and one that clones your repo and installs it from source.
  - Make sure to clearly state this difference in your installation instructions, and explain the difference between them.
  - Have the first option just list the `pip install pystarshade` option as that's sufficient for this case.
  - For the second option, provide the cloning instructions as you already to, and tell your users to install the package from source, in editable mode, so that any changes they do in the code is actually reflected at runtime. The command for this is (while in the top-level repository directory): `pip install -e .`

### Output data management

2. Implement a different default output path that separates the data from the code. I can suggest one of two options:
  - Instead of defaulting to inside the package location, default to a folder in the user's home directory that your code can create when this is done for the first time.
  - Alternatively, you could require an environment variable that the user needs to set with their desired default output path. In this case, you would need to specify this requirement in the installation instructions and also update your code base to use that environment variable.

### Dependencies

3. Simplify things and make `astropy` and `hcipy` normal dependencies. It takes time to figure out currently when I would want to install them and when not, and there is no disadvantage to adding them to the required dependencies (I think). In this way you eliminate a minor headache.
4. Include a Python requirement of at least 3.12 to your package.
  - The module `pkg_resources` is not part of the Python standard library anymore since Python 3.12. This means that you need to list `setuptools` (which the module is a part of) as an explicit dependencies for Python versions >= 3.12. The simplest way is to bump your Python requirement to >= 3.12 and add `setuptools` to your dependency list.
  - The reason you haven't run into this issue before is because `hcipy` installs `setuptools` as part of its installation. However, all direct dependencies of your own package should always be listed.
  - For more info, see https://github.com/mu-editor/mu/issues/2485
