### 2025-04-10 xiaziyna/PyStarshade#9 @ivalaginja
More information on example scripts
This is part of https://github.com/openjournals/joss-reviews/issues/7917

It's super nice to have this many example scripts! I just find it a little hard to identify where would be a good place to start, and what script aims to do what. Below some change requests and suggestions.

1. Have you considered putting your examples into notebook format instead? I will not insist on this as part of this review, but since you work a lot with visual data, it can be immensely helpful to be able to plot things inside a notebook instead of having to open figures from files saved to disk. Also, notebooks can be much easier included in the documentation, if you want to go that route in the future.
2. Whether you keep them as scripts or convert to notebooks, please provide more info inside of them:
  a) What is the concrete example script doing
  b) Does it require the example data or not
  c) How long is it expected to run (because some of them really take some time...)
