### 2025-04-10 xiaziyna/PyStarshade#10 @ivalaginja
Remove documentation redundancy
This is part of https://github.com/openjournals/joss-reviews/issues/7917

1. You have redundant information between the readme and the documentation. I would recommend to consolidate, and keep everything inside the documentation only. If not, there is an increased risk of having diverging information which could lead to confusion.

What I would keep/put in the Readme:
- Very short overview, shorter than you have now.
- Link to documentation, prominently displayed
- Contributing
- License

What I would move to/consolidate with docs:
- Installation instructions and dependency info
- Quickstart and examples

2. It would also be nice to have an overview of your research/science/instrumentation case in the documentation - a brief overview of what is a starshade, how does it work, what is it used for; maybe with one figure. This doesn’t need to be longer than 2-3 paragraphs, but it would be very helpful as it explains what your code is used for in the first place. And then you can follow with all info about your code, which is currently in the overview section and reads well (both in the readme and in the docs, see comment above).
