### 2025-04-10 xiaziyna/PyStarshade#7 @Jashcraf
I believe the conventional definition is something like the following:

$$
T = \frac{\sum_i PSF(x_i, y_i)}{\sum_i EP(x_i, y_i)}
$$

Where $T$ is throughput, $PSF$ is your point-spread function at the image plane, and $EP$ is your entrance pupil function. Core throughput is the same quantity with the $PSF$ being multiplied by some small aperture (with a $0.7 \lambda/D$ radius or something). Is this how you are computing throughput? (or something equivalent)

This is straightforward for internal occulters (coronagraphs) - but this metric doesn't work in as straightforward a fashion for external occulters because the Starshade is modifying the $EP$. The hangup I have with the integrated throughput $>1$ is because it implies that Starshades create energy, which seems physically impossible to me.

Both in the reference you linked and in the HabEx final mission report, they show similar figures

Though the Habex analysis elects to ignore this bump (see page 6-19 of the Habex final report)
> Interestingly enough the starshade exhibits a small amplification near the tips. However, for the purpose of contrast specification, this gain is ignored and the throughput unity in the wider field is used...

In many of Stuart Shaklan's recent works - it looks like they drop integrated throughput as a metric and instead compute Geometric throughput (essentially, 1-apodization)
source: https://www.spiedigitallibrary.org/conference-proceedings-of-spie/13092/130925D/Starshade-concepts-for-the-Habitable-Worlds-Observatory/10.1117/12.3020798.full

Regardless, I think I'm satisfied with the implementation in the code, as long as the normalization is consistent with the literature.

Might be worth including this demo in the docs and saying that this phenomenon is due to the definition of throughput - and not necessarily because starshades create energy? Or perhaps because the starshade is at some particular Fresnel distance it results in constructive interference at the IWA.
