+++ README.md
@@ -10,8 +10,6 @@ PyStarshade is a Python library for Starshade (or any external occulter) simulat
 
 What is a starshade? A starshade is a particular apodization (mask), which flown at Fresnel distances from a telescope achieves star-light suppression for imaging exoplanets in orbit around the star. 
 
-When a starshade is aligned with a star, starlight is diffracted by the starshade $s(\mathbf{x})$ onto a telescope aperture. For a starshade mask $s(u, v)$ which is zero inside the mask and unity outside, at a wavelength $\lambda$ and starshade-telescope distance $z$, the field at the telescope aperture is $f_{\lambda}(x, y)$ is related to the Fourier transform of the starshade mask:
- $$f_{\lambda}[u, v] \propto \mathcal{F} \left( s(u, v) e^{\frac{j \pi}{\lambda z} (u^2 + v^2)} \right) \left[ \frac{x}{\lambda z} ,\frac{y}{\lambda z} \right]$$
 Numerical diffraction calculations must use a very small numerical resolution $d u$ of the starshade $s(u, v)$ in order to accurately calculate starlight suppression. Using a standard FFT to perform these calculations is inefficient as very large zero-padding factors are needed to sample the field at the telescope aperture. The Bluestein FFT is a technique to calculate arbitrary spectral samples of a propagated field, indirectly using FFTs and therefore benefiting from their efficiency. For an $N \cdot N$ starshade mask, and an $M \cdot M$ telescope aperture, the Bluestein FFT approach achieves a complexity of $O((N+M)^2 \log (M+N))$. This technique is utilized in multiple aspects of the optical train to efficiently propagate fields.
 
 
@@ -31,11 +29,12 @@ You can install PyStarshade using pip:
 pip install pystarshade
 ```
 
-To use pre-computed data, use [git lfs](https://git-lfs.com): 
+Or to use pre-computed data, use [git lfs](https://git-lfs.com): 
 ```bash
 $ git clone https://github.com/xiaziyna/PyStarshade.git PyStarshade
 $ cd PyStarshade
 $ git lfs pull
+pip install -e .
 ```
 
 ## Dependencies
