+++ docs/content/background.rst
@@ -1,0 +1,65 @@
+Starshade Basics
+----------------
+
+What is a Starshade?
+~~~~~~~~~~~~~~~~~~~
+
+A **starshade** is a large, petal-shaped spacecraft designed to work in tandem with a space telescope for the direct imaging of exoplanets. Acting as an external occulter, it flies in precise formation with the telescope, positioned tens of thousands of kilometers away, to block the bright light from a star while allowing the faint light from orbiting planets to reach the telescope's aperture.
+
+How Does a Starshade Work?
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The starshade creates a shadow by suppressing starlight through diffraction. Its petal-like design minimizes light scattering around the edges, producing a dark region where the telescope can observe. When aligned with a target star, the starshade reduces the star's light by a factor of :math:`10^{-10}` to :math:`10^{-11}` as demonstrated in subscale tests, enabling the detection of planets that are billions of times fainter. The starshade can maintain this contrast over a broad wavelength range, from ultraviolet (UV) to near-infrared (NIR). Thereby allowing for the detection and characterization of planetary atmospheres, including the search for biosignatures such as water vapor (H₂O), oxygen (O₂), carbon dioxide (CO₂), methane (CH₄), and ozone (O₃). Furthermore, starlight is supppressed prior to light entering the telescope aperture, a starshade maintains a high level of optical throughput (close to or around 100%). In PyStarshade you can simulate throughput for different apertures. 
+
+
+Figure: Starshade Concept
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. image:: starshade.png
+   :alt: Starshade Diagram
+
+*Caption:* A starshade blocks starlight, casting a shadow on the telescope, allowing planet light to be imaged.  
+
+For more on starshade science, see resources like `Mennesson et al. (2024) <https://doi.org/10.1117/1.JATIS.10.3.035004>`_ or `Shaklan et al. (2023) <https://doi.org/10.1117/12.2677793>`_.
+
+----
+
+Optical Model Overview
+---------------------
+
+PyStarshade is built upon a Fourier optics model that simulates the propagation of light from an exoplanetary scene through a starshade and telescope system. The model consists of two main stages of optical propagation:
+
+1. **Starshade to Telescope Aperture (Fresnel Diffraction):**  
+   Light from the exoplanetary scene, modeled as a collection of point sources, propagates to the starshade. The starshade mask diffracts the on-axis light, and this diffracted field is calculated at the telescope aperture using Fresnel diffraction. For an on-axis source, the diffracted field at wavlength :math:`\lambda` is :math:`f_{P,\lambda}^0` is computed as:
+   
+   .. math::
+      
+      f_{P,\lambda}^0[x, y] = \frac{e^{\frac{i\pi}{\lambda z}(x^2 + y^2)}}{\lambda z} \mathcal{F}\left( s[\zeta, \eta] e^{\frac{i\pi}{\lambda z}(\zeta^2 + \eta^2)} \right) \left[ \frac{x}{\lambda z}, \frac{y}{\lambda z} \right]
+   
[…]
