+++ docs/index.rst
@@ -56,6 +56,7 @@ Contents
 --------
 
 .. toctree::
+    content/background
     content/usage
     content/fft
 
