+++ setup.py
@@ -10,7 +10,7 @@ setup(
     packages=find_packages(),
     install_requires=[
         "numpy",
-        "scipy",
+        "scipy"
     ],
     extras_require={
         "extra": [
@@ -30,5 +30,6 @@ setup(
         "Programming Language :: Python :: 3.9",
     ],
     python_requires=">=3.6"
+
 )
 
