+++ docs/content/usage.rst
@@ -28,6 +28,7 @@ If you wanted to design your own drm (or modify one of these), go to the file 'd
         'iwa': [30 / 9.52e+07 / mas_to_rad]  # R_ss / dist_ss_t in MAS
     }
 
+The parameter ``grey_mask_dx`` describes the pixel sampling in physical units of a mask file. For example, ``01m`` describes Δs = 0.01 m. 
 The easiest way to interface with PyStarshade is via the StarshadeProp class. 
 
 .. code-block:: python
@@ -51,10 +52,21 @@ Note: there is an important setting called 'chunk', if this is invoked a chunked
 
 
 
-Fourier propagator
+Dimensions
 -------------------
 
 In PyStarshade, the classes and functions are generally designed to act on a spatial grid of size N * N, where N is an odd number and the origin is at (N/2, N/2). 
+As an example, calling the function 'source_field_to_ccd', this function
+takes as input a 2D source-field of size (N_s, N_s) and spatial sampling ds and returns the 2D output
+field incident on a CCD of size (N_pix, N_pix) and pixel size dp. 
+
+.. code-block:: python
+
+    from pystarshade.simulate_field import source_field_to_ccd
+
+    source_field_to_ccd(source_field, wl, dist_xo_ss, dist_ss_t, focal_length_lens, radius_lens, 
+                            N_s = 333, N_x = 6401, N_t = 1001, N_pix = 4001, 
+                            ds = 10*0.03*au_to_meter, dx = 0.01, dt = 0.0116, dp=.5*1.9e-7)
 
 Data
 ----------
@@ -75,24 +87,10 @@ The data directory is structured like so:
 
 If you have new masks for the starshade, or telescope aperture masks, place them in the correct folders (starshade_masks and pupils respectively). 
 
+Location of Generated Files
+--------------------------
 
-Usage
---------
-
-The simplest way to use PyStarshade is by using the precomputed pupil fields and the StarshadeProp class as described. The StarshadeProp class is designed to abstract away sampling calculations, as well as pre-compute data products and interface with them.
-However, if desired, one may more directly interface with the optical propagation classes (Fresnel and Fraunhofer), or at a slightly higher level, the functions 'pupil_to_ccd' and 'source_field_to_pupil'.
-
-As an example, calling the function 'source_field_to_ccd', this function
-takes as input a 2D source-field of size (N_s, N_s) and spatial sampling ds and returns the 2D output
-field incident on a CCD of size (N_pix, N_pix) and pixel size dp. 
-
-.. code-block:: python
-
-    from pystarshade.simulate_field import source_field_to_ccd
-
-    source_field_to_ccd(source_field, wl, dist_xo_ss, dist_ss_t, focal_length_lens, radius_lens, 
-                            N_s = 333, N_x = 6401, N_t = 1001, N_pix = 4001, 
-                            ds = 10*0.03*au_to_meter, dx = 0.01, dt = 0.0116, dp=.5*1.9e-7)
+The generated pupil fields and PSF basis files are stored in the ``data/fields`` and ``data/psf`` directories, respectively, allowing for reuse in future simulations without repeating the computationally intensive propagation steps.
 
 Please see the examples folder for detailed examples!
 
@@ -103,3 +101,47 @@ PyStarshade can take as input any pixelized source-field such as Haystacks model
 (so far a point source and Gaussian source). If you wish to perform propagation using analytic descriptions, please 
 use 'pystarshade.simulate_field.point_source_to_ccd'. 
 
+Starshade Masks
+---------------
+
+The starshade mask itself is not generated by PyStarshade, we have collected masks from various sources (SISTERS/diffraq) and interpolated/upsampled these to grey-scale masks which are stored in the ``data/masks/starshade_masks`` directory. 
+
+Usage
+--------
+
[…]
