+++ docs/requirements.txt
@@ -1,5 +1,7 @@
 sphinx==7.1.2
 sphinx-rtd-theme==1.3.0rc1
 numpy
+nbsphinx
+ipykernel
 scipy
 hcipy
