+++ examples/simple_point_source.py
@@ -1,6 +1,8 @@
 from pystarshade.diffraction.util import mas_to_rad, au_to_meter, pc_to_meter, data_file_path
 from pystarshade.propagator import StarshadeProp
 
+#Creates a blank 1001×1001 field with two delta-like sources, initializes an HWO starshade with a hex pupil, propagates those point sources through the starshade at 500 nm, and writes the output intensity to test_field.npz. 
+
 hwo_starshade = StarshadeProp(drm = 'hwo')
 hwo_starshade.gen_pupil_field()
 hwo_starshade.gen_psf_basis('hex')
