+++ docs/content/usage.rst
@@ -88,24 +88,21 @@ The data directory is structured like so:
 If you have new masks for the starshade, or telescope aperture masks, place them in the correct folders (starshade_masks and pupils respectively). 
 
 Location of Generated Files
-===============
-
+^^^^^^^^^^^^^^^
 
 The generated pupil fields and PSF basis files are stored in the ``data/fields`` and ``data/psf`` directories, respectively, allowing for reuse in future simulations without repeating the computationally intensive propagation steps.
 
 Please see the examples folder for detailed examples!
 
 Input data
-===============
-
+^^^^^^^^^^^^^^
 
 PyStarshade can take as input any pixelized source-field such as Haystacks model or an ExoVista model, or analytic descriptions of sources
 (so far a point source and Gaussian source). If you wish to perform propagation using analytic descriptions, please 
 use 'pystarshade.simulate_field.point_source_to_ccd'. 
 
 Starshade Masks
-===============
-
+^^^^^^^^^^^^^^^
 
 The starshade mask itself is not generated by PyStarshade, we have collected masks from various sources (SISTERS/diffraq) and interpolated/upsampled these to grey-scale masks which are stored in the ``data/masks/starshade_masks`` directory. 
 
@@ -116,16 +113,14 @@ The simplest way to use PyStarshade is by using the precomputed pupil fields and
 
 
 Using Chunked FFT for Large Masks
-===============
-
+^^^^^^^^^^^^^^^
 
 **Important:** By default when ``chunk=1`` when calling ``gen_pupil_field(chunk = 1)``, the mask will be propagated in chunks. To use this, the generated starshade mask must be a memmap ``.dat`` filetype. You can generate a memmap file by running the ``make_memmap`` script inside the ``masks`` directory on one of the existing masks. We do not include masks generated in this format as they occupy a large disk space. Set ``chunk=0`` to use an npz file instead - beware you may run out of memory. 
 
 Chunked FFT processing avoids RAM limitations when working with very large masks. The computation processes the starshade mask of size (N × N) in chunks of size (N//N_chunk × N//N_chunk). By default, N_chunk is set to 4 in ``pystarshade.simulate_field.source_field_to_pupil``.
 
 Full PSF Generation Workflow
-===============
-
+^^^^^^^^^^^^^^^
 
 Here's a complete example of generating and using a PSF basis:
 
