+++ docs/index.rst
@@ -44,7 +44,6 @@ Contents
 .. toctree::
     content/background
     content/usage
-    content/fft
     content/solar_system.ipynb
 
 API
