+++ docs/paper/paper.md
@@ -41,7 +41,7 @@ Starshades are external occulting masks designed to suppress starlight by a fact
 
 ![Simulated imaging of a synthetic exoscene (ExoVista) with three visible exoplanets at a wavelength of 500 nm. A 60 m starshade configuration and a 6m segmented pupil was used for this example. \label{fig:example}](exo_scene.png){ width=50% }
 
-Aimed at being broadly useful for numerically intense Fourier optical simulations, tools are provided for efficient Fresnel and Fraunhofer propagation using Fourier spectral sampling with Bluestein Fast Fourier Transforms (FFT). Furthermore, a novel technique to chunk these FFT calculations is implemented and mitigates memory bottlenecks. A set of unit tests are provided to validate the Fourier diffraction tools. 'PyStarshade' pre-computes fields, point-spread-functions and throughput grids for different stages of the optical propagation chain, allowing for flexibility in modifying instrument parameters or telescope aperture masks. 'PyStarshade' optionally interfaces with HCIPy [@hcipy] to generate telescope apertures and the resulting imaging in a streamlined fashion. 
+Aimed at being broadly useful for numerically intense Fourier optical simulations, tools are provided for efficient Fresnel and Fraunhofer propagation using Fourier spectral sampling with Bluestein Fast Fourier Transforms (FFT). Furthermore, a novel technique to chunk these FFT calculations is implemented and mitigates memory bottlenecks. A set of unit tests are provided to validate the Fourier diffraction tools. 'PyStarshade' pre-computes fields, point-spread-functions and throughput grids for different stages of the optical propagation chain, allowing for flexibility in modifying instrument parameters or telescope aperture masks. 'PyStarshade' optionally interfaces with HCIPy [@hcipy] to generate telescope apertures and the resulting imaging. 
 
 # Related software
 The Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance (`SISTER`) [@sisters:2022] implemented in Matlab performs detailed end-to-end simulations of starshade imaging, further including instrument noise sources such as solar glint, and tools for modeling an exoplanetary system and background objects. The `SISTER` tool performs diffraction calculations from starshade to telescope aperture using the boundary diffraction wave method due to [@cady]. `Fresnaq` [@fresnaq] is a Matlab code to efficiently compute diffraction of a starshade at the telescope aperture via the non-uniform FFT [@barnett]. `Diffraq` [@diffraq] is a Python implementation of `Fresnaq`, this toolbox also provides an implementation of the boundary diffraction wave method and tools to generate starshade masks [@harness].  The Exoplanet
