### 2025-04-23 xiaziyna/PyStarshade#10 @xiaziyna
Thanks for the comments!

1. I have cut down some of the lengthy descriptions about the package. I have a strong preference for keeping an example output, install instructions and a quickstart in the README. I think this is quite typical among small scientific software tools (e.g. https://www.pyopensci.org/python-package-guide/tutorials/add-readme.html) and helps the user to quickly get started using a package without needing to look through wordy documentation.

2. I have added in several new sections to the documentation to provide extra background and info - https://pystarshade.readthedocs.io/en/latest/content/background.html (starshade basics, optical model overview and starshade concepts).
