+++ docs/content/usage.rst
@@ -11,7 +11,7 @@ To start pick a starshade configuration, some options provided are ('wfirst', 'h
 
     drm = 'hwo'
 
-If you wanted to design your own drm (or modify one of these), go to the file 'data/drm.py' and change some the instrument parameters. The drm parameters define the size of the starshade, it's flight distance, number of petals, wavelength bands, size of the telescope aperture and focal length. In the drm is a parameter called 'dx_' which is the pixel size of the starshade mask in meters. You need to have a starshade mask on file (pystarshade cannot generate these from scratch). Some masks are provided for the drm's listed (stored in data/masks). An example drm is:
+If you wanted to design your own drm (or modify one of these), go to the file 'data/drm.py' and change some the instrument parameters. The drm parameters define the size of the starshade, it's flight distance, number of petals, wavelength bands, size of the telescope aperture and focal length. An example drm is:
 
 .. code-block:: python
 
@@ -104,12 +104,13 @@ use 'pystarshade.simulate_field.point_source_to_ccd'.
 Starshade Masks
 ^^^^^^^^^^^^^^^
 
-The starshade mask itself is not generated by PyStarshade, we have collected masks from various sources (SISTERS/diffraq) and interpolated/upsampled these to grey-scale masks which are stored in the ``data/masks/starshade_masks`` directory. 
+The starshade mask itself is not generated by PyStarshade. Some masks are provided for the drm's listed (stored in data/masks), collected from various sources (SISTERS/diffraq) and interpolated/upsampled these to grey-scale masks which are stored in the ``data/masks/starshade_masks`` directory. If you have a starshade mask defined by a set of edge-points (locii) and want to generate a grey-scale (antialiased) mask for use with PyStarshade, run the script `mp_gen_mask.py` to do so. If you want to generate diffracted fields using a chunked FFT (the diffraction from the mask is processed in chunks), then generate a memory mapped mask using `make_memmap.py`. See further down for more info.
 
 Usage
 --------
 
-The simplest way to use PyStarshade is by using the precomputed pupil fields and the StarshadeProp class as described. The StarshadeProp class is designed to abstract away sampling calculations, as well as pre-compute data products and interface with them. Pystarshade utilizes a PSF basis to simulate imaging. Part of the flow in generating this basis involves simulating a diffracted field incident on the telescope aperture. This field needs only be generated once and can be used to generate different PSF basis for different apertures. Diffracted fields at the telescope aperture live in the `data/fields`` directory. A set of pre-generated fields for the HWO starshade drm are available with the git-lfs install. 
+The simplest way to use PyStarshade is by using the precomputed pupil fields and the StarshadeProp class as described. The StarshadeProp class is designed to abstract away sampling calculations, as well as pre-compute data products and interface with them. Pystarshade utilizes a PSF basis to simulate imaging. 
+Generating the diffracted pupil fields from scratch for different wavelengths can be quite compute intensive depending on the size of the starshade and may take several hours. However, this field needs only be generated once and can be used to generate different PSF basis for different apertures. Diffracted fields at the telescope aperture live in the `data/fields`` directory. A set of pre-generated fields for the HWO starshade drm are available with the git-lfs install. 
 
 
 Using Chunked FFT for Large Masks
@@ -117,7 +118,8 @@ Using Chunked FFT for Large Masks
 
[…]
