+++ docs/content/usage.rst
@@ -106,7 +106,7 @@ Starshade Masks
 
 The starshade mask itself is not generated by PyStarshade. Some masks are provided for the drm's listed (stored in data/masks), collected from various sources (SISTERS/diffraq) and interpolated/upsampled these to grey-scale masks which are stored in the ``data/masks/starshade_masks`` directory. If you have a starshade mask defined by a set of edge-points (locii) and want to generate a grey-scale (antialiased) mask for use with PyStarshade, run the script `mp_gen_mask.py` to do so. If you want to generate diffracted fields using a chunked FFT (the diffraction from the mask is processed in chunks), then generate a memory mapped mask using `make_memmap.py`. See further down for more info.
 
-Usage
+Generating a PSF Basis
 --------
 
 The simplest way to use PyStarshade is by using the precomputed pupil fields and the StarshadeProp class as described. The StarshadeProp class is designed to abstract away sampling calculations, as well as pre-compute data products and interface with them. Pystarshade utilizes a PSF basis to simulate imaging - however only a pre-computed PSF basis at a single wavelength is included.  
