### 2025-04-24 xiaziyna/PyStarshade#9 @xiaziyna
1. I have created a new notebook to merge some of the existing examples and provide an overview of the package. https://pystarshade.readthedocs.io/en/latest/content/solar_system.html

2. I have added a short comment at the top of each script describing it's purpose. Most of the example scripts use some portion of either pre-computed data, or data provided from ExoVista exoplanetary scenes. I have added a line to the install instructions to recommend performing the git LFS install in order to run the examples:

![Image](https://github.com/user-attachments/assets/df875592-966a-4083-8c72-0d07822627a9)

Besides the use of ExoVista data, and if the user has ran the appropriate script to format a starshade mask (documented in the Usage section), the example scripts should auto-generate the missing parts of the pipeline. Performing an end-to-end starshade simulation is quite computationally intensive so I have added some time estimates in the 'Usage /  Generating a PSF Basis' section of the documentation.
