### 2025-04-25 xiaziyna/PyStarshade#8 @xiaziyna
Hi,

Thanks for all the detail!

**PyPI installation vs. cloning the repo**
1. I have amended the LFS install instructions to pip install in editable mode. I have also revised the installation instructions to clearly separate these two install instructions - with warnings.
The diffraction tools/library certain can be used without the data. However, the user will need to supply data (such as a starshade mask) and depending on the size of the starshade/computational resources available, this could be computationally time-intensive to simulate from scratch. For this reason I have recommended the user install with git-LFS.

Output data management
2. The output data path is currently not being used by any script (it was created for personal use). I have deleted this path and will add in an external output path if any script in the future needs one.

Dependencies
3. I am hesitant to add HCIPy and Astropy as dependencies due to their limited use in the package/not wanting to install a bunch of extra packages on someones machine. 1) HCIPy is only used to generate example apertures using the 'apodization.pupil.make_pupil' script,  the user can also provide their own aperture (as is used in the example notebook). 2) Astropy is only used to open an ExoVista fits file in example scripts. So perhaps a solution would be to include further instructions in the installation guide? I will add these as dependencies if you think it is better to do so.

4. Thanks for pointing this out to me. I have updated the `setup.py` and added setuptools as a requirement. I believe this is always compatible with earlier Python 3 versions, so I have left in those versions for now.
