### 2025-04-30 xiaziyna/PyStarshade#10 @ivalaginja
@xiaziyna The new sections about background info are fantastic! As for the readme, this is always a personal choice. Your changes are fine by me, I will just throw in a word of caution to look out for maintenance of both, readme and docs, and divergence between the two as you continue working on and using your package.

I consider this solved so I will close the issue.
