### 2025-04-30 xiaziyna/PyStarshade#8 @ivalaginja
@xiaziyna looking a lot better!

1. The installation instructions are so much cleaner now, thanks! I will make sure to test this, but if everything works fine, nothing else to do here.
2. From what I saw, you use `plt.savefig()` and `np.savez_compressed()` extensively throughout the repo though, at least in the examples. These outputs need to be directed to somewhere that is not inside your repository. What is your current output data management plan for these examples?
3. If your packages has dependencies, then you should indeed list them as such, even if they are used "rarely" - even a single usage justifies this. Otherwise you are providing an incomplete installation. Users need to be able to manage their local installations themselves, so cluttering their machines is also not a good reason to withhold dependencies. On the contrary, the more streamlined your installation process, the easier it will be for your users. As a side note, nowadays people who run Python *should* be using virtual environment solutions anyway (e.g. conda or others), so installations for new packages and projects would be isolated enough that your installation would not be invasive into the user's setup at all. You cannot really dictate this part though, and this often remains up to the user.
4. Indeed, this is backwards-compatibe so you should be fine. However, now you list support for Python versions 3, 3.6-3.9 and 3.12, wich makes it look like 3.10 and 3.11 are unsupported. Is this the case? If you support in fact everything starting from 3.6 and above, it might be less confusing if you just list ` "Programming Language :: Python :: 3"`, since you list `python_requires=">=3.6"  # Keep supporting 3.6 and above` anyway.
