### 2025-05-01 xiaziyna/PyStarshade#6 @Jashcraf
This is a superb demo!
