### 2025-05-02 87111cd
output directory management/setup
