+++ setup.py
@@ -11,25 +11,15 @@ setup(
     install_requires=[
         "numpy",
         "scipy",
+        "hcipy",
+        "astropy",
         "setuptools"  # Add setuptools as a dependency for all Python versions
     ],
-    extras_require={
-        "extra": [
-            "hcipy",
-            "astropy"
-        ]
-    },
     classifiers=[
         "Development Status :: 2 - Pre-Alpha",
         "Intended Audience :: Science/Research",
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Operating System :: OS Independent",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.12",  # Add this to indicate 3.12 support
     ],
-    python_requires=">=3.6"  # Keep supporting 3.6 and above
 )
