### 2025-05-02 d57ea02
output directory management/setup
