### 2025-05-02 xiaziyna/PyStarshade#8 @xiaziyna
Thanks for taking the time to be thorough in your explanations and providing expertise!

2. True. I have added an output directory set to the users home: '~/PyStarshade' (set in the pystarshade/config.py file and instantiated when called for the first time). I maintain an internal data directory (for the files used by pystarshade, starshade masks, optical fields etc.). Is this ok?

3.4. Your explanation makes a lot of sense and I have updated the setup to include the dependencies. I have also replaced all the random Python versions with Python 3.
