### 2025-05-04 xiaziyna/PyStarshade#8 @ivalaginja
Looks good! With the updated dependencies, there would just need to be an update to that section in the readme:
https://github.com/xiaziyna/PyStarshade/tree/main?tab=readme-ov-file#dependencies

And I just realized that your installation instructions (https://pystarshade.readthedocs.io/en/latest/#installation) should also give an indication of the dependencies, ideally with instructions on how to install them. Could I ask you do fix these two things please? You're almost there!
