### 2025-05-04 xiaziyna/PyStarshade#8 @ivalaginja
And about the output details in comment/question 2.: This looks like a great solution! If I run into any issues while testing this, I can easily open a new issue so I will consider this done here!
