### 2025-05-05 xiaziyna/PyStarshade#8 @xiaziyna
Hi! I have updated the README to list the updated dependencies.
