### 2025-05-05 xiaziyna/PyStarshade#11 @ivalaginja
Suggestion: Add `.gitignore` file to repo
Result of review in https://github.com/openjournals/joss-reviews/issues/7917.

When I install your package locally from source, I get all the build files dumped inside the repo (which is normal). Since these are local files that shouldn't be committed and pushed, I recommend including a `.gitignore` file in the top level of the repo. You can fill it with:
- a filter for local build files (i.e., `*.egg-info/`)
- a filter for pycache files (i.e., `.pyc`)
- file extensions for the most common code editors (e.g., `.iml` for PyCharm, `.vs` for VSCode)
- randomly created hidden files based on the OS (e.g., `.DS_Store` for MacOS)

And example for a very simple `.gitignore` file is:
```
*.pyc
*.iml
*.egg-info/
.ipynb_checkpoints/
.idea/
.vs
.DS_Store
```

In this way they don't pollute your git working tree when working with a clone of the repo directly. This is handy both for developers (you) and your users.
