### 2025-05-05 xiaziyna/PyStarshade#9 @ivalaginja
The rework of the examples are great, and I love the notebook to get started with! Thank you!
