### 2025-05-05 xiaziyna/PyStarshade#12 @ivalaginja
`pytest` needed as dependency/requirement
This is part of https://github.com/openjournals/joss-reviews/issues/7917.

Since you use `pytest` as your testing suite, this needs to be added to your dependency list and requirements.

Example import here:
https://github.com/xiaziyna/PyStarshade/blob/1af86013a936f70028cef77e6492e8ba7fedb941/tests/test_bluestein_fft.py#L6
