### 2025-05-05 xiaziyna/PyStarshade#14 @ivalaginja
Requested documentation cleanup
This is part of https://github.com/openjournals/joss-reviews/issues/7917.

You have all you need in the documentation, but I believe that it can be somewhat cleaned up to significantly improve its reading. Below some requests/ideas on how to achieve that:

1. The general documentation sections render ok in the menu, but their ordering, and how they show up or don’t in the left menu vs. on a single page is a bit confusing. Hare are some suggestions on how to improve on that.
    - I find it weird to have separate sections in the menu on the left, while some of them go to a new page and some don’t. E.g., the menu item “Starshade Basics” opens a new page, even with two (menu-clickable) subsections. The menu items “Optical Model Overview”, “Bluestein FFT” and “Starshade Concepts” all have their own, top-level menu items, but they’re all on the same page, while the menu item “Usage” opens a new page with its own subsections that are now indicated as such in the menu. Whichever way you choose, please rearrange this in a consistent way.
    - Documentation landing page
        - I would put the current documentation landing page into its own menu item to the left, and make its sub-components (Installation, Contents, etc.) clickable subsections (like you have in “Usage” for example). This is something you want people to find easily no matter which part of your documentation they have open at any given moment. Clicking the “PyStarshade” thingy on top is of course an option, but less intuitive.
        - On this first page, maybe put “Contents” and “API” last. They are long subsections and its easy to miss the license and contact sections since it’s less likely that people scroll past that.
        - Add contributing guidelines to the documentation, maybe to this landing page, it seems to fit in here well. You have an encouragement right now (only in the Readme), it is usually easier for users if you explicitly tell them what they should do if they want to contribute. E.g.: “If you want to contribute new code, feel free to directly open a pull request. If you want to discuss a new feature or ask for help in implementing them, please go ahead and open an issue.” → “Feel free to reach out” is very vague, since it’s not like people can send you a message over Github, and then they need to know if you want them to send you an email, or open an issue, etc.
2. In the Usage section (https://pystarshade.readthedocs.io/en/latest/content/usage.html#usage), it would be nice to have a top-level overview of what it means to run your code. What I mean here are some general bullet points: “What can you do with PyStarshade?” which is currently only covered with the first sencence: “PyStarshade will compute fields and PSF files for a given starshade configuration and store these on disk for simulating imaging of new scenes.”
    - Expand on this sentence. You can refer to your nice figure in the starshade description! Something like, “PyStarshade allows you to compute the diffracted fields and PSFs in planes A, B, C and D of your starshade configuration, see Fig. XX.”
    - And then give a list of inputs you need to do those computations, e.g. “To do so PyStarshade requires”:
        - A starshade configuration given by starshade mask, telescope aperture, …
        - An exoplanet scence given by XYC, which is usually loaded from the ExoVista package.
    - → I probably forgot a billion things here, but I hope it is clear what I mean!
3. The last code example in your Overview section (https://pystarshade.readthedocs.io/en/latest/content/usage.html#overview) is uncommented, but the actual problem is that the parameter `source_field` is undefined when first used. Not sure if this snipped of code is here by accident, or if it’s just missing docs and/or the parameter definition?
4. Same comment like above for the code snipped in the Dimensions section (https://pystarshade.readthedocs.io/en/latest/content/usage.html#dimensions)
5. Aha, it seems like both of the above are actually listed at the very bottom of this page (https://pystarshade.readthedocs.io/en/latest/content/usage.html#full-psf-generation-workflow). It looks like they can be deleted above then? In any case, even down here, `source_field` remains undefined and the code those unrunnable. Could you edit this example to make it runnable?
6. In your basis section (https://pystarshade.readthedocs.io/en/latest/content/usage.html#generating-a-psf-basis) I would put the sentence: “Generating the diffracted pupil fields from scratch for different wavelengths can be quite compute intensive depending on the size of the starshade and may take several hours.” into a warning box or similar, to make sure it is highlighted. Same for: “Computing the PSF basis itself may take several minutes or up to an hour, depending on the bandwidth, spectral sampling and source-field sampling.”
7. Is the `out` folder in your Data section (https://pystarshade.readthedocs.io/en/latest/content/usage.html#data) still correct with your updated output directory into the users home directory?
