### 2025-05-05 xiaziyna/PyStarshade#15 @ivalaginja
Requested example notebook cleanup
I love the new notebook! It makes it immediately clear what your code is doing, and how it's doing it 👍

There are a couple of minor kinks that will need to be ironed out, but overall this looks very good already.

1. PyStarshade optical summary
    - The labels in the figure are “Source field0”,  “Transmission function 1”, etc., while the text talks about “point sources/exoplanet scene”, “starshade mask”, “telescope” etc. Please add the denominations you use in the text to the figure, otherwise it is very hard for a non-expert to consolidate the two.
    - It might also help to be explicit and indicate the focal planes and focal planes specifically.
    - Are the “far-field” and “near-field” labels all correct in your figure? The middle one seems ok, but to my understanding, the Fraunhofer propagation with the focal length f on the right-hand side is a far-field propagation (which is done through Fraunhofer diffraction/Fourier transforms). In the same way,  propagation over a specific distance d1 (on the left-hand side in your figure) is then Fresnel propagation/near-field propagation. Or am I interpreting something wrong in the figure?
2. Starshade masks
    - You provide a direct link to `ExoVista` later on, can you do the same for `SISTERS` and `diffraq`?
3. PSF basis
    - Ok to not run the code to calculate a diffracted field in the notebook - can you give a reason why though? I suspect it takes a long time - can you add a number that tells us how long it runs on a specific machine? Say I am on a MacBook Pro (13-inch, M1, 2020) running MacOS 12.2 (Monterey), I could quote one single time for this cell. Can you provide this for your machine so the user gets a feel for it?
4. Simulating imaging
    - Minor: In your text you say you’ll use an off-axis telescope, while in the code cell the on-axis version is used by default.
    - Running this cell also takes a while, a preemptive time estimate would be neat.
    - Hm this is running really long… is there a way to just show the code that would run this, with a time estimate (like in the earlier example in PSF basis), but you use a ore-computed PSF basis in the notebook itself?
    - Showing the aperture before the scene generation might actually be more logical, unless there is some code sequence that I am not seeing that prevents this. Since you also use the language: “Let’s plot this” and then you proceed to plot the focal intensity, I presume there is just an issue with cell ordering right now 🙂
5. I will have to do the rest of the notebook at a later time since I am stuck at the cell with `gen_scene()`...
