### 2025-05-05 #7917 @ivalaginja
Hi @xiaziyna, thanks for addressing all my previous issues! I am now continuing to move through the checklist. In doing so, I have identified a couple more things to fix, which I review in the issues below. Sorry that this is so iterative, but I believe it might be easier if I send you what I have already to give you time to have a look while I continue my review.

As always, if something is unclear or doesn't seem to make sense, please do reach out!

- https://github.com/xiaziyna/PyStarshade/issues/11
- https://github.com/xiaziyna/PyStarshade/issues/12
- https://github.com/xiaziyna/PyStarshade/issues/13
- https://github.com/xiaziyna/PyStarshade/issues/14
- https://github.com/xiaziyna/PyStarshade/issues/15
