### 2025-05-06 xiaziyna/PyStarshade#13 @xiaziyna
Thanks for catching this.

This test is testing computing an output chunk of a DFT broken into 4 quadrants of space - it failed here because when computed with the FFT comparison case the zero-padded size N_X was less than 2*N_out (hence not enough samples to break it into 4 output quadrants and the slicing failed). I have corrected this and the tests should run successfully. I will consider setting up a Github Actions Workflow for the tests at some point in the future!
