### 2025-05-06 xiaziyna/PyStarshade#13 @ivalaginja
It's fixed indeed! 👍
