### 2025-05-06 xiaziyna/PyStarshade#12 @ivalaginja
Checked and it installs it too now. Thanks!
