### 2025-05-16 xiaziyna/PyStarshade#7 @Jashcraf
Sounds like a great plan to me! I'll close this issue since it's been a while - my apologies!
