### 2025-05-16 xiaziyna/PyStarshade#16 @Jashcraf
More information on the Bluestein FFT in paper.md
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

The paper could use some additional information on the Bluestein FFT in the software paper. The section in the readthedocs is very good (linked below), and something similar would be very helpful for the software paper

https://pystarshade.readthedocs.io/en/latest/content/background.html#bluesteins-fft

It would also be helpful for readers to have the connection made to the chirp-z transform in the paper. Correct me if I'm wrong, but my understanding is that the Bluestein FFT is an application of the CZT?
https://en.wikipedia.org/wiki/Chirp_Z-transform#Bluestein's_algorithm

If you are interested in a side-by-side comparison, `prysm` has the CZT implemented for similar calculations
https://github.com/brandondube/prysm/blob/d45826f11dc173786bb621d90904bc020caf3102/prysm/fttools.py#L423
