### 2025-05-16 xiaziyna/PyStarshade#17 @Jashcraf
Community Guidelines in Documentation
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

JOSS requests that packages post **Community Guidelines**:

Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I saw some contribution guidelines in the README.md, but the documentation website needs a dedicated page for community guidelines. One good place to start is the Contributor Covenant, if the authors are interested:

https://www.contributor-covenant.org/
