### 2025-05-16 xiaziyna/PyStarshade#18 @Jashcraf
Instructions on running automated tests in documentation
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7917

The documentation could use some information on the landing page for installing + running tests to verify functionality. Nothing particularly involved needed, just a way to walk through the process of testing your installation of Pystarshade. Even if it's just a note under Installation directing new users to running

```
pytest tests/
```

Another great addition for Pystarshade's continued development would be adding a Github action for automated testing. I didn't see one in the repository's Actions, so I'm just noting it here. If the authors are interested, I use codecov: https://about.codecov.io/, but you are free to choose your own method of running automated tests.
