### 2025-05-16 #7917 @Jashcraf
@ivastar @xiaziyna Just posted a couple of issues on the Pystarshade repository after running through the JOSS review criteria. Nothing major - kudos to @xiaziyna this is looking really great!
