+++ docs/index.rst
@@ -21,13 +21,30 @@ License and Citation
 
     Taaki, Kamalabadi, Kemball. PyStarshade: simulating high-contrast imaging of exoplanets with starshades
 
-
 Table of Contents
 -----------------
 
 .. toctree::
     :maxdepth: 2
-    content/install       Installation Guide
-    content/background    Background and Theory
-    content/usage         Usage Instructions
+    content/install 
+    content/background
+    content/usage
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Tutorials
+
+   content/solar_system   Simulating the solar system (Jupyter notebook)
+
+.. toctree::
+   :maxdepth: 1
+   :caption: API Reference
+
+   content/test           Core classes & functions reference
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contributing
+
+   content/contribute     How to report issues, request features, or submit code
 
