+++ docs/content/background.rst
@@ -1,3 +1,6 @@
+Background
+==========
+
 What is a Starshade?
 --------------------
 
