+++ docs/content/test.rst
@@ -1,6 +1,8 @@
+PyStarshade API Reference
+-------------------------
+
 .. autoclass:: pystarshade.propagator.StarshadeProp
    :members: set_mission_params, gen_pupil_field, gen_pupil, gen_psf_basis, gen_scene
-
 
 .. autoclass:: pystarshade.diffraction.diffract.Fresnel
     :members: 
