+++ docs/content/usage.rst
@@ -1,5 +1,5 @@
 Usage
-=========================
+=====
 
 Overview
 ----------
