+++ docs/index.rst
@@ -40,6 +40,5 @@ Table of Contents
 
    content/solar_system
    content/contribute
-
-    `API Reference <content/test.html>`_
+   API Reference <content/test.html>
 
