+++ docs/content/usage.rst
@@ -106,29 +106,6 @@ The generated pupil fields and PSF basis files are stored in the ``data/fields``
 
 Please see the examples folder for detailed examples!
 
-Input data
-^^^^^^^^^^^^^^
-
-To perform these computations, PyStarshade requires the following inputs:
-
-- **Starshade configuration (Design Reference Mission, DRM)**: A dictionary specifying instrument parameters, including:
-
-  - Starshade mask parameters: radius (:math:`R_{ss}`, e.g., 30 m for HWO), number of petals (e.g., 24), and pixel sampling (:math:`\Delta s`, e.g., 1 mm to achieve :math:`10^{-10}` contrast).
-  - Telescope parameters: aperture diameter (:math:`D_P`, e.g., 6 m), focal length (:math:`f`, e.g., 10 m), and aperture mask (e.g., segmented on-axis or off-axis designs).
-  - Optical parameters: wavelength band (:math:`\lambda`, e.g., 500–1000 nm), flight distance between starshade and telescope (:math:`d_{s \rightarrow P}`, e.g., :math:`9.52 \times 10^7` m), and Fresnel number (:math:`f_{ss} = \frac{R_{ss}^2}{\lambda d_{s \rightarrow P}}`, e.g., 9.5–18.9).
-  - Sampling parameters: starshade mask sampling (:math:`\Delta s`), telescope aperture sampling (:math:`\Delta P`, e.g., 2 cm), and focal plane sampling (:math:`\Delta f`, e.g., 2 mas).
-- **Starshade mask file**: a binary starshade mask on a grid.
-- **Telescope aperture mask**: A binary or grayscale mask representing the telescope aperture (e.g., HWO’s segmented on-axis or off-axis designs). These can be user-defined or loaded from HCIPy.
-- **Exoplanet scene (optional for imaging)**: A flux distribution of the astrophysical scene. PyStarshade can take as input any pixelized source-field such as Haystacks model or an ExoVista model, or analytic descriptions of sources
-(so far a point source and Gaussian source). If you wish to perform propagation using analytic descriptions, please 
-use 'pystarshade.simulate_field.point_source_to_ccd'. 
-
-- **Computational settings**: Optional parameters to control numerical precision, such as:
-
-  - Chunking factor (:math:`N_{chunk}`) for memory-efficient Bluestein FFT (B-FFT) computations.
-  - Wavelength sampling (:math:`\Delta \lambda`, e.g., 50 nm steps).
-  - PSF basis grid size (e.g., 150 × 150 terms for HWO).
-
 
 Starshade Masks
 ^^^^^^^^^^^^^^^
@@ -154,6 +131,23 @@ A set of pre-generated pupil fields for the HWO starshade drm are available with
 PyStarshade does not assume any azimuthal symmetry and will compute a fully unique PSF for each source pixel.  Once computed, the PSF basis can be used to simulate imaging for different scenes.
 
 Choices for sampling and resolution in the optical pipeline can be modified, however we have generally chosen defaults that opt for precision over speed and memory usage. 
+
+Input data
+^^^^^^^^^^^^^^
+
+To perform these computations, PyStarshade requires the following inputs:
+
+- **Starshade configuration (Design Reference Mission, DRM)**: A dictionary specifying instrument parameters, including:
+
+  - Starshade mask parameters: radius (:math:`R_{ss}`, e.g., 30 m for HWO), number of petals (e.g., 24), and pixel sampling (:math:`\Delta s`, e.g., 1 mm to achieve :math:`10^{-10}` contrast).
[…]
