+++ docs/content/usage.rst
@@ -8,10 +8,17 @@ PyStarshade will compute fields and PSF files for a given starshade configuratio
 
 - **Diffracted field at telescope**: A single diffracted field is computed per wavelength in the specified band, assuming an on-axis source.
 - **A PSF basis** in the focal plane for a grid of source positions, enabling imaging simulations of exoplanetary scenes. The PSF basis is computed from the telescope field by masking regions with the chosen telescope aperture mask and propagating the field to the focal plane (described in Aime et al. 2024 and Taaki et al. 2024). 
-- **Imaging simulations**: PSF's are scaled by source flux of an astrophysical scene (input by the user) and co-added.
+- **Imaging simulations**: PSF's are scaled by source flux of an astrophysical scene (input by the user) and co-added to simulate imaging.
 
-Note - the optical propagation tools can be applied generally for Fresnel/Fraunhofer diffractive simulations.
+Note - the optical propagation tools can be applied more generally for Fresnel/Fraunhofer diffractive simulations, or for use of 2D spectral sampling. 
 
+Dimensions
+-------------------
+
+In PyStarshade, the classes and functions are generally designed to act on a spatial grid of size N * N, where N is an odd number and the origin is at (N/2, N/2). 
+As an example, calling the function 'source_field_to_ccd', this function
+takes as input a 2D source-field of size (N_s, N_s) and spatial sampling ds and returns the 2D output
+field incident on a CCD of size (N_pix, N_pix) and pixel size dp. 
 
 Configuring the starshade
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -55,30 +62,6 @@ After running this code, a field incident on the telescope aperture will be stor
 Note: there is an important setting called 'chunk', if this is invoked a chunked FFT will be used to compute the fields using a memory mapped starshade mask. This avoids hitting RAM bottlenecks, but you will need to have enough disk memory for the memmap starshade mask of :math:`\left( \frac{diameter_{ss}}{dx} \right)^2` floats. By default, this chunked FFT will process the starshade mask of size (N * N) in chunks of size (N/N_chunk * N/N_chunk). To modify the N_chunk parameter change it inside 'pystarshade.simulate_field.source_field_to_pupil', by default N_chunk is set to 4.
 
 
-.. code-block:: python
-
-    pupil_type = 'hex'
-    hwo_starshade.gen_psf_basis(pupil_type = pupil_type)
-    focal_intensity = hwo_starshade.gen_scene(pupil_type, source_field.astype(np.float32), 500e-9)
-
-
-
-Dimensions
--------------------
-
-In PyStarshade, the classes and functions are generally designed to act on a spatial grid of size N * N, where N is an odd number and the origin is at (N/2, N/2). 
-As an example, calling the function 'source_field_to_ccd', this function
-takes as input a 2D source-field of size (N_s, N_s) and spatial sampling ds and returns the 2D output
-field incident on a CCD of size (N_pix, N_pix) and pixel size dp. 
-
-.. code-block:: python
-
-    from pystarshade.simulate_field import source_field_to_ccd
-
-    source_field_to_ccd(source_field, wl, dist_xo_ss, dist_ss_t, focal_length_lens, radius_lens, 
-                            N_s = 333, N_x = 6401, N_t = 1001, N_pix = 4001, 
-                            ds = 10*0.03*au_to_meter, dx = 0.01, dt = 0.0116, dp=.5*1.9e-7)
-
 Data
 ----------
 
@@ -112,7 +95,7 @@ Starshade Masks
 
[…]
