### 2025-06-03 xiaziyna/PyStarshade#14 @xiaziyna
Hi - thanks for the feedback. I have focused on cleaning up the documentation.
1.

![Image](https://github.com/user-attachments/assets/7204eeb1-25d4-4255-a725-dd38827b3589)

For the contributing guidelines I added a stand alone page.

2. I have added a top level overview of the general uses. I have also added a section under the 'Generating and using a PSF Basis' section of 'Input Data', summarizing what the user needs to collect in order to simulate imaging.
3. I have deleted the initial instance of this.
4. I have removed this code-snippet as it isn't needed here. This section was just meant to describe the centering of the fields.
5. Corrected the snippet including 'source_field' in the 'full PSF generation' section to include a toy source field model of a star + planet.
6. I have added these warnings
7. I have updated this and added a line specifying 'New data generated from examples is located at the top level: 'PyStarshade/out'.'
