### 2025-06-03 xiaziyna/PyStarshade#17 @xiaziyna
Hi!

Thanks for the suggestion.
I have added a dedicated page for contributing with clearer instructions: https://pystarshade.readthedocs.io/en/latest/content/contribute.html
