### 2025-06-03 xiaziyna/PyStarshade#19 @ivalaginja
JOSS paper review
This relates to https://github.com/openjournals/joss-reviews/issues/7917.

You have come a long way with the software itself and its documentation, I see no major blockers on that side for a publication with JOSS, in particular once all open issues have been addressed.

For the paper, there are a couple of major comments I have that would need addressing before we can proceed. These concern all sections of your paper:

### Summary

The summary is intended to provide a top-level overview of your software, and the scientific context it folds into. In particular, it should provide enough information so that a "diverse, non-specialist audience" can capture what type of scientific question the software is used for.

Currently, the summary just states what a starshade is, and that it's used to image faint exoplanets, all in one sentence. Then you jump to some exoplanet background, albeit in a very convoluted way, and then back to what PyStarshade does.

I find this very hard to understand in terms of global context. I only realized now, because I am myself in this field and this section initially seemed minimalist but sufficient. However, I believe that the following points need to be mentioned in your paper summary in order to give a valid representation of the field:

- Direct imaging as a tool for exoplanet science. Why is direct imaging used, to answer what kind of scientific questions. What has been done with direct imaging so far, what are the current limitations of this method.
- What are the key elements in the field of the field of direct imaging. Starlight suppression can typically be done either with internal or external occulters, so I would argue you need to mention both to be able to position starshades alongside coronagraphs.
- What are past, current and anticipated missions that thought of using a starshade instrument, what's their status.
- What does PyStarshade do, which you have already in there, and what is it used for.

The top-down flow of the above gives the aforementioned non-specialist audience a picture of where your particular software plugs into that is currently missing. Note that it does not have to be particularly long, half a page would be more than enough. But it would be important to hit all of the above points.

### Statement of need, and Related software

Your current statement of need gives a description of what PyStarshade does. This is definitely necessary, but not quite sufficient to fully meet the paper criteria for this section. You give a nice explanation of what the software calculates and *how*, but what's missing is the *why* - "what problems the software is designed to solve" - and the *who* - "who the target audience is". Essentially, you should explain to us which type of researcher could use your software to solve which type of research questions.

On the same topic, you do have a section describing related software. What is missing though is how that other software relates to yours, what are the commonalities and differences ("how this software compares to other commonly-used packages")? Do some of them use other optical propagation techniques? Are they for specific instruments and yours is not? I do believe you have most necessary info about these other packages in there, but what's missing is the concrete relation to your package. This is probably also the place where you can highlight one of the unique points of PyStarshade, which is the use of Bluestein FFTs, which @Jashcraf also pointed out (https://github.com/xiaziyna/PyStarshade/issues/16).

Whether you keep the current separation of sections is totally up to you, you're the one who will see best what makes most sense once the changes are in place.

### References

Your reference list is a bit thin. How come most of the references you have in the docs, which are very good choices, haven't made it in the paper? In particular in an updated Summary section like I describe above, you will want to include at least one or two major references to coronagraphy, and then even more on basic starshade physics and concepts - in the end, this is what your software is all about! The references inside Shaklan et al. (2024), one of the papers in your docs, is a good point to start. You don't need to add 10 starshade paper references, but at least one of two general concept papers would be important to have here, as well as starshade mission concept papers.

Note how the Barnett et al paper has a published version in JATIS, which should be cited in favor of the non-peer-reviewed arXiv preprint.
