### 2025-06-03 #7917 @xiaziyna
Hi Iva! Thanks for all the comments, I want to get round to addressing them one by one and am also working through some of the existing issues.

Just to answer your general notes briefly:
The telescope apertures were provided by Rhonda Morgan (private communication 2024), who told me they were baselined for HWO. I haven't been able to get a hold of her or I would ask for further clarification. I will put a note about these as plausible/reasonable designs (segmented/and obscured), but not confirmed. Currently I do not know where I can access the EAC designs. The HWO [starshade](https://www.spiedigitallibrary.org/conference-proceedings-of-spie/12680/1268010/A-starshade-for-NASAs-habitable-worlds-observatory-concept--exoplanet/10.1117/12.2677793.short) is a point design.

I will try and correct this or mention WFIRST is now called Roman. The rendezvous starshade design was not adopted, so the only mention of this starshade concept is in the [original proposal](https://ui.adsabs.harvard.edu/abs/2018AAS...23112109S/abstract) where Roman was called WFIRST (one reason to retain some of the WFIRST notation).

A smaller D = 10 m starshade would operate at a different (shorter) flight distance to maintain a Fresnel number ~ 10 - 20 (generally at this Fresnel number a fixed contrast is maintained). A starshade works on the principle of Fresnel diffraction to cancel the spot of Arago that would be present with a circular diffracting mask to reach a contrast level < 10^-10.
At a fixed wavelength the radius of the dark shadow scales approximately as the radius of the starshade, a smaller starshade requires a smaller telescope. If flight distance is decrased, the IWA ~ R_s / z will be larger, thus the starshade is limited in its ability to see in the Habitable zone.

For a flight distance z = 1e8m, the starshade would be closer to the HWO point design with D=60 m and a Fresnel number between 10-20 over the 500-1000 nm band.

There is a JATIS Special Section on Starshades with this nice figure illustrating the design space:

![Image](https://github.com/user-attachments/assets/a0d81bae-8462-4315-8640-e39b84a48899)

The smallest starshade design is the [Mdot](https://slab.stanford.edu/projects/mdot) 3m design (but the goal was not to image exoplanets in the hab zone).
