### 2025-06-03 xiaziyna/PyStarshade#15 @xiaziyna
Modifications to the example notebook:

1. I have modified the figure to more clearly reflect the text, i.e. 'starshade mask', 'telescope aperture', 'pupil plane' and 'focal plane'.
So the propagation from telescope aperture to focal field is near-field diffraction, however reduces to computing the Fraunhofer diffraction of the pupil mask. The reason is as follows: a simple telescope model treats the aperture as a thin-lens, the effective pupil is the pupil mask multiplied by a phase delay function (to account for the varying phase delay of light traveling through the lens). When you compute the Fresnel diffraction of the phase delay * the pupil mask, you obtain the Fraunhofer diffraction of the pupil-mask alone. The Fresnel number is 'near-field' but the focused field is the Fourier transform of the incident field*pupil mask. Introduction to Fourier Optics by Goodman has a description of this. Alternatively [here](https://web.mit.edu/2.710/Fall06/2.710-wk10-a-ho.pdf) are some slides.

2. I now include a link to SISTER and diffraq - actually the SISTER software link appears down, so I have added a link to the software article for the time being.

3. I have added time estimates based on my machine (Thinkpad P14s) in the notebook for the PSF basis (~ 15 min) and the imaging (~ 1 min). The diffracted fields do take a while ~ 1 hr to compute on my more powerful machine. I have added a description of this to the documentation. This could be sped up but is not worth the premature optimization at this point.

4. Corrected 'off-axis' to 'on-axis'. The problem with including a pre-computed PSF basis is it generates more data the user should download. For my own work I am using very high spatial sampling with no symmetry assumptions, at all stages of optical propagation and this is the default in the library, I could change these default samplings and size of fields to make the PSF generation a bit quicker but then the results will diverge from my work. Since the notebook is rendered in the documentation - is it not sufficient for the user to just look here to see the complete output?
I have rearranged the cells to show the aperture before the imaging generation - but note I show the scene loading before the aperture because I want to keep all operations related to generating the PSF basis co-located and the user needs to declare a propagator object before loading the aperture/setting the mask.
