+++ docs/content/install.rst
@@ -21,3 +21,16 @@ Instead install the package from source, in editable mode and use `git lfs <http
     git lfs pull
     pip install -e .
 
+
+Testing
+^^^^^^^
+
+To run the tests, navigate to the root directory of the **pystarshade** repository and execute:
+
+.. code-block:: bash
+
+   pytest tests/
+
+This will automatically discover and run all unit tests in the `tests/` directory.
+
+If all tests pass, your installation is working correctly. If a test fails - please open an issue!
