### 2025-06-03 xiaziyna/PyStarshade#18 @xiaziyna
Thanks for pointing this out!

I have added some instructions to the installation page on running unit tests. Plan to add automated testing in the future.
