### 2025-06-03 xiaziyna/PyStarshade#19 @xiaziyna
Thank you for your detailed comments.

I have a paper I am planning to submit to AAS journals within a few days on starshade simulations/optics. However, I am worried about potential overlap/self plagarism.
The paper focuses on describing an optical model of a starshade and performing simulations of core throughput. When fine-scale features are present in the telescope aperture, it is necessary to use large zero-padding factors to accomplish Fourier optics calculations with FFT's, rendering them intractable. The Bluestein FFT allows one to move across various resolutions/scales without the fixed input-output constraints of a standard FFT. The paper describes the computational complexities of various standard optical calculation methods, and describes why the Bluestein FFT is more efficient in this setting - and points out where exactly they offer a benefit. At the time I started working on this, there were not good implementations of this technique for optics (in Python), so that was part of my motivation in writing this tool.

Before I proceed with changing the paper, could you give me some clarification on the dual [AAS submission](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing) policy?

I've heard as an approximate guideline, that less than 15% of content can overlap between publications - I think a description of direct imaging as you outlined would not overlap, nor would a general description of the B-FFT (which is in the other paper, however is a general method), and a more detailed description of existing software utilities, however I am trying to keep the 'software' paper somewhat distinct from 'starshade optics/computational techniques' so knowing whether these should or should not overlap is helpful!
