### 2025-06-04 xiaziyna/PyStarshade#19 @ivalaginja
Hey @xiaziyna, very exciting to hear about your upcoming paper!

You are right to be thinking about this. To be totally sure, I will be pinging your assigned editor over in the review PR, because they will known the full details.

What I can tell you is that you gave a pretty good distinction between the two, the "scientific concept paper", and the "software paper" yourself above. All of the nitty-gritty technicalities should definitely go into your AAS journal paper. With JOSS covering the software itself, the JOSS paper should really just be providing enough information to understand the general scientific context in which the associated software finds its application. I usually think of the JOSS paper as a condensed version of the introduction section of the science concept paper, with specific focus on the software. This would mean that all important points are mentioned and put in relation to each other, without fully describing them. I am pretty sure that this is also what Jaren was aiming for with his issue. No need for a full explication of the Bluestein FFTs, just mention why you chose them, what problem they solve, and then I believe you can even refer to the science concept paper for further details.
