### 2025-06-04 xiaziyna/PyStarshade#19 @xiaziyna
Thanks @ivalaginja, that definitely helps clarify! I will get on with these updates.
