+++ docs/paper/paper.bib
@@ -9,11 +9,124 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@book{direct,
+  title={Direct imaging of exoplanets},
+  author={Traub, Wesley A and Oppenheimer, Ben R},
+  year={2010},
+  publisher={University of Arizona Press, Tucson}
+}
+
+
+@article{Follette_2023,
+doi = {10.1088/1538-3873/aceb31},
+url = {https://dx.doi.org/10.1088/1538-3873/aceb31},
+year = {2023},
+month = {sep},
+publisher = {The Astronomical Society of the Pacific},
+volume = {135},
+number = {1051},
+pages = {093001},
+author = {Follette, Katherine B.},
+title = {An Introduction to High Contrast Differential Imaging of Exoplanets and Disks},
+journal = {Publications of the Astronomical Society of the Pacific},
+abstract = {This tutorial is an introduction to High-Contrast Imaging, a technique that enables astronomers to isolate light from faint planets and/or circumstellar disks that would otherwise be lost amidst the light of their host stars. Although technically challenging, high-contrast imaging allows for direct characterization of the properties of circumstellar sources. The intent of the article is to provide newcomers to the field a general overview of the terminology, observational considerations, data reduction strategies, and analysis techniques high-contrast imagers employ to identify, vet, and characterize planet and disk candidates.}
+}
+
+@BOOK{NAP26141,
+  author    = "National Academies of Sciences, Engineering, and Medicine",
+  title     = "Pathways to Discovery in Astronomy and Astrophysics for the 2020s",
+  isbn      = "978-0-309-46734-6",
+  doi       = "10.17226/26141",
+  abstract  = "We live in a time of extraordinary discovery and progress in astronomy and astrophysics. The next decade will transform our understanding of the universe and humanity's place in it.  Every decade the U.S. agencies that provide primary federal funding for astronomy and astrophysics request a survey to assess the status of, and opportunities for the Nation's efforts to forward our understanding of the cosmos. Pathways to Discovery in Astronomy and Astrophysics for the 2020s identifies the most compelling science goals and presents an ambitious program of ground- and space-based activities for future investment in the next decade and beyond.  The decadal survey identifies three important science themes for the next decade aimed at investigating Earth-like extrasolar planets, the most energetic processes in the universe, and the evolution of galaxies.  The Astro2020 report also recommends critical near-term actions to support the foundations of the profession as well as the technologies and tools needed to carry out the science.",
+  url       = "https://nap.nationalacademies.org/catalog/26141/pathways-to-discovery-in-astronomy-and-astrophysics-for-the-2020s",
+  year      = 2023,
+  publisher = "The National Academies Press",
+  address   = "Washington, DC"
+}
+
+@ARTICLE{bluestein,
+  author={Bluestein, L.},
+  journal={IEEE Transactions on Audio and Electroacoustics}, 
[…]
