+++ docs/paper/paper.md
@@ -49,7 +49,7 @@ Fourier optics simulations of starshades and telescope apertures require fine sp
 Furthermore, a novel technique to chunk Bluestein FFT calculations is implemented and mitigates memory bottlenecks. A set of unit tests are provided to validate the Fourier diffraction tools. `PyStarshade` pre-computes fields, point-spread-functions and throughput grids for different stages of the optical propagation chain, allowing for flexibility in modifying instrument parameters or telescope aperture masks. `PyStarshade` optionally interfaces with HCIPy [@hcipy] to generate telescope apertures and the resulting imaging. 
 
 # Related software
-The Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance (`SISTER`) [@sisters:2022] implemented in Matlab performs detailed end-to-end simulations of starshade imaging, further including instrument noise sources such as solar glint, and tools for modeling an exoplanetary system and background objects. The optical simulation pipeline of `PyStarshade` follows a similar approach to that of SISTER, employing a precomputed PSF basis for image simulation. However, `SISTER` computes diffraction from the starshade to the telescope aperture using the boundary diffraction wave method due to [@cady], performing a 1D integral along the starshade's edge, followed by a DFT to obtain the final PSF basis. Several codes focus on calculating the starshade diffraction incident on the telescope aperture; `Fresnaq` [@fresnaq] is a Matlab code to efficiently compute diffraction of a starshade via the non-uniform FFT [@barnett]. `Diffraq` [@diffraq] is a Python implementation of `Fresnaq`, this toolbox also provides an implementation of the boundary diffraction wave method and tools to generate starshade masks [@harness]. Starshade masks used in `PyStarshade` are obtained from `SISTER`, or were generated with `Diffraq`. Outside of imaging simulations, the Exoplanet Open-Source Imaging Mission Simulator (`EXOSIMS`) [@exosims] provides methods for orbital simulation of starshades and fuel consumption calculations.
+The Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance (`SISTER`) [@sisters:2022] implemented in Matlab performs detailed end-to-end simulations of starshade imaging, further including instrument noise sources such as solar glint, and tools for modeling an exoplanetary system and background objects. The optical simulation pipeline of `PyStarshade` follows a similar approach to that of `SISTER`, employing a precomputed PSF basis for image simulation. However, `SISTER` computes diffraction from the starshade to the telescope aperture using the boundary diffraction wave method due to [@cady], performing a 1D integral along the starshade's edge, followed by a DFT to obtain the final PSF basis. Several codes focus on calculating the starshade diffraction incident on the telescope aperture; `Fresnaq` [@fresnaq] is a Matlab code to efficiently compute diffraction of a starshade via the non-uniform FFT [@barnett]. `Diffraq` [@diffraq] is a Python implementation of `Fresnaq`, this toolbox also provides an implementation of the boundary diffraction wave method and tools to generate starshade masks [@harness]. Starshade masks used in `PyStarshade` are obtained from `SISTER`, or were generated with `Diffraq`. Outside of imaging simulations, the Exoplanet Open-Source Imaging Mission Simulator (`EXOSIMS`) [@exosims] provides methods for orbital simulation of starshades and fuel consumption calculations.
 
 # Acknowledgements
 
