+++ docs/paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 A key challenge in detecting exoplanets and characterizing their atmospheres, is distinguishing their faint reflected or emitted light from the star they orbit. Direct-imaging addresses this by suppressing starlight to reveal orbiting exoplanets [@direct]. Direct imaging can be achieved through two primary approaches: internal occulters, coronagraphs, which use optics within the telescope to block starlight, and external occulters, starshades, petal shaped masks some 10's of meters across flown in tandem with a telescope. Current direct imaging efforts, primarily using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory, have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, to see a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel]. 
 
-While a starshade has not yet been flown in space, in subscale demonstrations a starshade suppresses starlight by a factor of $10^{-10}$ or more, across a broad range of wavelengths needed to obtain measurements of water, oxygen, methane among other biogases [@mennesson]. Beyond suppressing starlight lie specific questions about the properties and configurations of exoplanetary systems that can be imaged, and which instrument designs are most suitable. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
+While a starshade has not yet been flown in space, in subscale demonstrations a starshade suppresses starlight by a factor of $10^{-10}$ or more, across the broad wavelength range necessary for detecting water, oxygen, methane, and other biomarkers [@mennesson]. Beyond starlight suppression, critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
 Several starshade concepts have been proposed but not adopted, including a 26 m diameter design to rendezvous with the Nancy Grace Roman Space Telescope  (a coronagraph scheduled for launch in the mid 2020's) [@ngrst_rendezvous]. A 52 m design as part of the Habex concept [@habex]. NASA’s Habitable Worlds Observatory (HWO), is in the early design phase with the goal to find and characterize a handful of earth-like exoplanets [@NAP26141]. A 60 m starshade operating over visible to infrared wavelengths (500–1000 nm), and a 35 m UV (250-500 nm) starshade, have been proposed for HWO [@hwo_ss].
 
@@ -46,7 +46,7 @@ Several starshade concepts have been proposed but not adopted, including a 26 m 
 
 Fourier optics simulations of starshades and telescope apertures require fine spatial pixel scales to accurately model diffraction without aliasing errors. Traditional Fast Fourier Transforms (FFTs) necessitate large zero-padding factors to accomplish this, increasing computational cost. `PyStarshade` implements the Bluestein FFT [@bluestein] to compute optical fields at arbitrary spatial scales, without zero-padding and therefore with greater efficiency than the FFT or discrete Fourier transform (DFT). In the associated paper submitted to AAS, we use `PyStarshade` to assess the complementary role of a 60 m HWO starshade [@hwo_ss] paired with 6 m segmented and obscured telescope apertures featuring cm-scale details. We report core throughput—the fraction of exoplanet light recovered—, a key metric governing exposure times needed to image an exoplanet and the overall yield of a mission.
 
[…]
