### 2025-06-04 xiaziyna/PyStarshade#16 @xiaziyna
Hi @Jashcraf,

Thanks for the feedback! I have revised the paper to include a slightly longer description of the Bluestein FFT. As mentioned in Issue #19, I will save a more lengthy discussion for my AAS paper.

![Image](https://github.com/user-attachments/assets/62e6af24-3842-4bbf-b6fe-83b95f8cc7f8)

[Here](https://web.ece.ucsb.edu/Faculty/Rabiner/ece259/Reprints/350_CZT%20article.pdf) is an article by Rabiner (the author of the CZT) about the history of the CZT/Bluestein FFT, it seems like the Bluestein FFT came first and the CZT was an extension to the z-transform. I have modified the paper to mention that the CZT is related!

Note I most often see Bluestein FFT used in computing, where it is used to compute prime-length FFT's.

Thanks for the link to prysm, interested to check this out in more detail.
