+++ docs/paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-A key challenge in detecting exoplanets and characterizing their atmospheres, is distinguishing their faint reflected or emitted light from the star they orbit. Direct-imaging addresses this by suppressing starlight to reveal orbiting exoplanets [@direct]. Direct imaging can be achieved through two primary approaches: internal occulters, coronagraphs, which use optics within the telescope to block starlight, and external occulters, starshades, petal shaped masks some 10's of meters across flown in tandem with a telescope. Current direct imaging efforts, primarily using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory, have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, to see a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel]. 
+A key challenge in detecting exoplanets and characterizing their atmospheres, is distinguishing an exoplanets faint reflected or emitted light from that of the star they orbit. Direct-imaging addresses this by suppressing light from the star to reveal orbiting exoplanets [@direct]. Direct imaging can be achieved through two approaches: internal occulters, coronagraphs, which use optics within the telescope to block starlight, and external occulters, starshades, petal shaped masks some 10's of meters across flown in tandem with a telescope. Current direct imaging efforts using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory, have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, to see a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel]. 
 
 While a starshade has not yet been flown in space, in subscale demonstrations a starshade suppresses starlight by a factor of $10^{-10}$ or more, across the broad wavelength range necessary for detecting water, oxygen, methane, and other biomarkers [@mennesson]. Beyond starlight suppression, critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
