+++ docs/paper/paper.bib
@@ -17,6 +17,22 @@
 }
 
 
+@INPROCEEDINGS{subscale,
+       author = {{Harness}, A. and {Shaklan}, S. and {Kasdin}, N. and {Willems}, P. and {Balasubramanian}, K.},
+        title = "{Laboratory demonstration of 10$^{-10}$ contrast at the inner working angle of a sub-scale starshade}",
+    booktitle = {American Astronomical Society Meeting Abstracts \#235},
+         year = 2020,
+       series = {American Astronomical Society Meeting Abstracts},
+       volume = {235},
+        month = jan,
+          eid = {373.10},
+        pages = {373.10},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020AAS...23537310H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
 @article{Follette_2023,
 doi = {10.1088/1538-3873/aceb31},
 url = {https://dx.doi.org/10.1088/1538-3873/aceb31},
