+++ docs/paper/paper.bib
@@ -31,7 +31,19 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
+@article{special,
+author = {Jonathan W. Arenberg and Anthony D. Harness and Rebecca M. Jensen-Clem},
+title = {{Special Section on Starshades: Overview and a Dialogue}},
+volume = {7},
+journal = {Journal of Astronomical Telescopes, Instruments, and Systems},
+number = {2},
+publisher = {SPIE},
+pages = {021201},
+keywords = {starshades, exoplanets, starlight suppression, high contrast imaging, technology development, Telescopes, Exoplanets, Stars, Shadows, Coronagraphy, Diffraction, Planets, Optical instrument design, Apodization, Observatories},
+year = {2021},
+doi = {10.1117/1.JATIS.7.2.021201},
+URL = {https://doi.org/10.1117/1.JATIS.7.2.021201}
+}
 
 @article{Follette_2023,
 doi = {10.1088/1538-3873/aceb31},
