+++ docs/paper/paper.md
@@ -40,13 +40,13 @@ While a starshade has not yet been flown in space, in subscale demonstrations a 
 Several starshade concepts have been proposed but not adopted, including a 26 m diameter design to rendezvous with the Nancy Grace Roman Space Telescope  (a coronagraph scheduled for launch in the mid 2020's) [@ngrst_rendezvous]. A 52 m design as part of the Habex concept [@habex]. NASA’s Habitable Worlds Observatory (HWO), is in the early design phase with the goal to find and characterize a handful of earth-like exoplanets [@NAP26141]. A 60 m starshade operating over visible to infrared wavelengths (500–1000 nm), and a 35 m UV (250-500 nm) starshade, have been proposed for HWO [@hwo_ss].
 
 # Statement of need
-`PyStarshade` provides a toolbox for performing optical simulations from source to focal plane with a configurable starshade telescope design. Complex electric fields are calculated at three planes of propagation (starshade, telescope aperture and focal plane) using Fresnel or Fraunhofer diffraction formula where appropriate. These are used to pre-compute a spatially and spectrally varying point-spread-function (PSF) basis to simulate imaging. First-order imaging characteristics of a starshade can be determined from analytic relations which depend on the starshades size, telescope aperture size, wavelength and flight distance. `PyStarshade` allows for second order imaging characteristics to be studied, including simulations of imaging for a discretized exoplanetary scene of flux, varying starshade mask and telescope aperture mask to study throughput and post-processing methods. An example is provided on using spectral- and time-dependent scenes consisting of planets, star and dust-disk, generated with `ExoVista` [@Stark_2022], the output imaging is shown in \autoref{fig:example}. `PyStarshade` is intended to be flexible and efficient in studying i) exoplanet retrievals, and, ii) instrument design.
+`PyStarshade` provides a toolbox for performing optical simulations from source to focal plane with a configurable starshade telescope design. Complex electric fields are calculated at three planes of propagation (starshade, telescope aperture and focal plane) using Fresnel or Fraunhofer diffraction formula where appropriate. First-order imaging characteristics of a starshade can be determined from analytic relations which depend on the starshades size, telescope aperture size, wavelength and flight distance. `PyStarshade` allows for second order imaging characteristics to be studied, including simulations of imaging for a discretized exoplanetary scene of flux, varying starshade mask and telescope aperture mask to study throughput and post-processing methods. An example is provided on using spectral- and time-dependent scenes consisting of planets, star and dust-disk, generated with `ExoVista` [@Stark_2022], the output imaging is shown in \autoref{fig:example}. `PyStarshade` is intended to be flexible and efficient in studying i) exoplanet retrievals, and, ii) instrument design.
 
 ![Simulated imaging of a synthetic exoscene (ExoVista) with three visible exoplanets at a wavelength of 500 nm. A 60 m starshade configuration and a 6 m segmented pupil was used for this example. \label{fig:example}](exo_scene.png){ width=50% }
 
[…]
