### 2025-06-05 xiaziyna/PyStarshade#19 @xiaziyna
Hi,

Thanks for the detailed feedback. I have completed a revision of the paper to include the requested extra content.

I have expanded the summary section to touch on all the points you raised - I will say I could perhaps shorten the last paragraph summarizing some of the starshade concepts.

Under the statement of need I have added a new paragraph (second paragraph) and slightly modified the first/last. The new second paragraph includes a brief description of the Bluestein FFT, why it's needed here, and a description of the AAS paper.

I have added more details to the section on related software, essentially SISTER is the only tool for imaging simulations (the other tools listed are only for simulating the diffracted field incident on the telescope, which is an important step in the pipeline but not the whole thing). I have described the similarities in the overall optical pipeline, and the differences (namely the diffraction computational method).

You are right, the original reference list was lacking! I've updated this with references to starshade concepts/other references. I was operating on the assumption the paper should be barebones based on the example paper provided.

I've included some general references for direct imaging (if you have better/other suggestions here, please let me know)!

Thanks for catching the Barnett/JATIS citation correction, this has been corrected.
