### 2025-06-12 xiaziyna/PyStarshade#19 @ivalaginja
@xiaziyna this is excellent! I really like all your changes. I will get back to you shortly with a simple review on typos and formulations, but your revised paper is great!
