### 2025-06-13 xiaziyna/PyStarshade#15 @ivalaginja
@xiaziyna I am unsure what I am doing wrong. I let the cell with `hwo_starshade.gen_psf_basis(pupil_type = pupil_type)` run all night (~9 hours) and it did not finish executing on my MacOS 12.2 using an Apple M1 chip with 16 GB or RAM. What could be the issue?
