+++ docs/paper/paper.bib
@@ -1,13 +1,18 @@
-@software{sisters:2022,
+@ARTICLE{sisters:2022,
        author = {{Hildebrandt}, Sergi R. and {Shaklan}, Stuart B. and {Cady}, Eric J. and {Turnbull}, Margaret C.},
-        title = "{SISTER: Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance}",
- howpublished = {Astrophysics Source Code Library, record ascl:2203.001},
-         year = 2022,
-        month = mar,
-          eid = {ascl:2203.001},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ascl.soft03001H},
+        title = "{Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance}",
+      journal = {Journal of Astronomical Telescopes, Instruments, and Systems},
+     keywords = {starshade, exoplanets, imaging simulations},
+         year = 2021,
+        month = apr,
+       volume = {7},
+          eid = {021217},
+        pages = {021217},
+          doi = {10.1117/1.JATIS.7.2.021217},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021JATIS...7b1217H},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @book{direct,
   title={Direct imaging of exoplanets},
@@ -157,13 +162,13 @@ doi = {10.1117/1.JATIS.7.2.021211},
 URL = {https://doi.org/10.1117/1.JATIS.7.2.021211}
 }
 
-@misc{fresnaq,
-  author = {Barnett, Alex H},
-  title = {fresnaq: FRESNel Areal Quadrature},
-  year = {2020},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/ahbarnett/fresnaq}
+@software{fresnaq,
+  author  = {Barnett, Alex H.},
+  title   = {fresnaq: fast Fresnel scalar diffraction from binary occulters and apertures},
+  version = {5181d86},              % commit hash
+  date    = {2021-01-13},
+  url     = {https://github.com/ahbarnett/fresnaq},
+  note    = {GitHub repository, commit 5181d86; accessed 2025-06-13}
 }
 
 @book{goodman2005introduction,
@@ -211,14 +216,13 @@ URL = {https://doi.org/10.1117/1.JATIS.7.2.021211}
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
-@misc{diffraq,
-  author = {Harness, Anthony},
-  title = {Diffraq: Diffraction with Areal Quadrature},
-  year = {2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/harnessa/diffraq}
+@software{diffraq,
+  author  = {Harness, Anthony},
+  title   = {diffraq: Diffraction with Areal Quadrature},
+  version = {a61613a},            % commit hash (Jul 26 2021)
+  year    = {2021},
+  url     = {https://github.com/harnessa/diffraq/tree/a61613a},
+  note    = {GitHub repository, commit a61613a; accessed 2025-06-13}
 }
 
 @article{Stark_2022,
