+++ docs/paper/paper.bib
@@ -14,27 +14,32 @@
 }
 
 
-@book{direct,
-  title={Direct imaging of exoplanets},
-  author={Traub, Wesley A and Oppenheimer, Ben R},
-  year={2010},
-  publisher={University of Arizona Press, Tucson}
-}
-
-
-@INPROCEEDINGS{subscale,
-       author = {{Harness}, A. and {Shaklan}, S. and {Kasdin}, N. and {Willems}, P. and {Balasubramanian}, K.},
-        title = "{Laboratory demonstration of 10$^{-10}$ contrast at the inner working angle of a sub-scale starshade}",
-    booktitle = {American Astronomical Society Meeting Abstracts \#235},
-         year = 2020,
-       series = {American Astronomical Society Meeting Abstracts},
-       volume = {235},
-        month = jan,
-          eid = {373.10},
-        pages = {373.10},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020AAS...23537310H},
+@INCOLLECTION{direct,
+       author = {{Traub}, W.~A. and {Oppenheimer}, B.~R.},
+        title = "{Direct Imaging of Exoplanets}",
+    booktitle = {Exoplanets},
+         year = 2010,
+       editor = {{Seager}, S.},
+        pages = {111-156},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010exop.book..111T},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@article{subscale,
+author = {Anthony Harness and Stuart Shaklan and Phillip Willems and N. Jeremy Kasdin and Bala K. Balasubramanian and Philip Dumont and Victor White and Karl Yee and Rich Muller and Michael Galvin},
+title = {{Optical verification experiments of sub-scale starshades}},
+volume = {7},
+journal = {Journal of Astronomical Telescopes, Instruments, and Systems},
+number = {2},
+publisher = {SPIE},
+pages = {021207},
+keywords = {starshades, optical model validation, high contrast imaging, exoplanet detection, vector diffraction, Diffraction, Design, Displays, Data modeling, Apodization, Geometrical optics, Polarization, Manufacturing, Cameras, Optics manufacturing},
+year = {2021},
+doi = {10.1117/1.JATIS.7.2.021207},
+URL = {https://doi.org/10.1117/1.JATIS.7.2.021207}
+}
+
 
 @article{special,
 author = {Jonathan W. Arenberg and Anthony D. Harness and Rebecca M. Jensen-Clem},
@@ -178,26 +183,29 @@ URL = {https://doi.org/10.1117/1.JATIS.7.2.021211}
   publisher={Roberts and Company Publishers}
 }
 
+
 @article{cady,
-  title={Boundary diffraction wave integrals for diffraction modeling of external occulters},
-  author={Cady, Eric},
-  journal={Optics Express},
-  volume={20},
-  number={14},
-  pages={15196--15208},
-  year={2012},
-  publisher={Optical Society of America}
+  author  = {Cady, Eric},
+  title   = {Boundary diffraction wave integrals for diffraction modeling of external occulters},
+  journal = {Optics Express},
+  volume  = {20},
+  number  = {14},
+  pages   = {15196--15208},
+  year    = {2012},
+  doi     = {10.1364/OE.20.015196},
+  url     = {https://doi.org/10.1364/OE.20.015196}
 }
 
 @article{harness,
-  title={Advances in edge diffraction algorithms},
[…]
