+++ docs/paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 A key challenge in detecting exoplanets and characterizing their atmospheres is distinguishing an exoplanets faint reflected or emitted light from that of the star they orbit. Direct-imaging addresses this by suppressing light from the star to reveal orbiting exoplanet's [@direct]. Direct imaging can be achieved through two approaches: internal occulters, called coronagraphs, which use optics within the telescope to block starlight, and external occulters, called starshades, petal-shaped masks some 10's of meters across in diameter, flown in tandem with a telescope [@special]. Current direct imaging efforts using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, seeing a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel] in flux ratio between star and planet". 
 
-While a starshade has not yet been flown in space, starshades achieve $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in subscale laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers like  water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
+While a starshade has not yet been flown in space, subscale starshades have demonstrated $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers like  water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
 Several starshade concepts have been proposed but not adopted, including a 26 m diameter design to rendezvous with the Nancy Grace Roman Space Telescope (a coronagraph scheduled for launch in the mid 2020's) [@ngrst_rendezvous] and a 52 m design as part of the Habex concept [@habex]. NASA’s Habitable Worlds Observatory (HWO), is in the early design phase with the goal to find and characterize a handful of Earth-like exoplanets [@NAP26141]. A 60 m starshade operating over visible to infrared wavelengths (500–1000 nm), and a 35 m UV (250-500 nm) starshade, have been proposed for HWO [@hwo_ss].
 
