+++ docs/paper/paper.bib
@@ -268,6 +268,13 @@ abstract = {Studies of future space- and ground-based exoplanet surveys often re
   note     = {Zenodo archive}
 }
 
+@unpublished{taaki_hwo_sim_2025,
+  author = {Taaki, Jamila S. and Kamalabadi, Farzad and Kemball, Athol J. and Corrales, Lia and Hero, Alfred O.},
+  title  = {Efficient Exoplanet Imaging Simulations of the Habitable Worlds Observatory},
+  year   = {2025},
+  note   = {in preparation},
+}
+
 @ARTICLE{habex,
        author = {{Gaudi}, B. Scott and {Seager}, Sara and {Mennesson}, Bertrand and {Kiessling}, Alina and {Warfield}, Keith and {Cahoy}, Kerri and {Clarke}, John T. and {Domagal-Goldman}, Shawn and {Feinberg}, Lee and {Guyon}, Olivier and {Kasdin}, Jeremy and {Mawet}, Dimitri and {Plavchan}, Peter and {Robinson}, Tyler and {Rogers}, Leslie and {Scowen}, Paul and {Somerville}, Rachel and {Stapelfeldt}, Karl and {Stark}, Christopher and {Stern}, Daniel and {Turnbull}, Margaret and {Amini}, Rashied and {Kuan}, Gary and {Martin}, Stefan and {Morgan}, Rhonda and {Redding}, David and {Stahl}, H. Philip and {Webb}, Ryan and {Alvarez-Salazar}, Oscar and {Arnold}, William L. and {Arya}, Manan and {Balasubramanian}, Bala and {Baysinger}, Mike and {Bell}, Ray and {Below}, Chris and {Benson}, Jonathan and {Blais}, Lindsey and {Booth}, Jeff and {Bourgeois}, Robert and {Bradford}, Case and {Brewer}, Alden and {Brooks}, Thomas and {Cady}, Eric and {Caldwell}, Mary and {Calvet}, Rob and {Carr}, Steven and {Chan}, Derek and {Cormarkovic}, Velibor and {Coste}, Keith and {Cox}, Charlie and {Danner}, Rolf and {Davis}, Jacqueline and {Dewell}, Larry and {Dorsett}, Lisa and {Dunn}, Daniel and {East}, Matthew and {Effinger}, Michael and {Eng}, Ron and {Freebury}, Greg and {Garcia}, Jay and {Gaskin}, Jonathan and {Greene}, Suzan and {Hennessy}, John and {Hilgemann}, Evan and {Hood}, Brad and {Holota}, Wolfgang and {Howe}, Scott and {Huang}, Pei and {Hull}, Tony and {Hunt}, Ron and {Hurd}, Kevin and {Johnson}, Sandra and {Kissil}, Andrew and {Knight}, Brent and {Kolenz}, Daniel and {Kraus}, Oliver and {Krist}, John and {Li}, Mary and {Lisman}, Doug and {Mandic}, Milan and {Mann}, John and {Marchen}, Luis and {Marrese-Reading}, Colleen and {McCready}, Jonathan and {McGown}, Jim and {Missun}, Jessica and {Miyaguchi}, Andrew and {Moore}, Bradley and {Nemati}, Bijan and {Nikzad}, Shouleh and {Nissen}, Joel and {Novicki}, Megan and {Perrine}, Todd and {Pineda}, Claudia and {Polanco}, Otto and {Putnam}, Dustin and {Qureshi}, Atif and {Richards}, Michael and {Eldorado Riggs}, A.~J. and {Rodgers}, Michael and {Rud}, Mike and {Saini}, Navtej and {Scalisi}, Dan and {Scharf}, Dan and {Schulz},
[…]
