+++ docs/paper/paper.md
@@ -44,7 +44,7 @@ Several starshade concepts have been proposed but not adopted, including a 26 m 
 
 ![Simulated imaging of a synthetic exoscene (ExoVista) with three visible exoplanets, as well as two planets within the region of starlight suppression, at a wavelength of 500 nm. The star to planet flux ratio of the visible planets range from $10^{-8}$ - $10^{-10}$. A 60 m HWO starshade configuration and a 6 m segmented pupil was used for this example. \label{fig:example}](exo_scene.png){ width=50% }
 
-Fourier optics simulations of starshades and telescope apertures require fine spatial pixel scales to accurately model diffraction without aliasing errors. Traditional Fast Fourier Transforms (FFTs) necessitate large zero-padding factors to accomplish this, increasing computational cost. `PyStarshade` implements the Bluestein FFT [@bluestein] (related to the chirp Z-transform) to compute optical fields at arbitrary spatial scales, without zero-padding and therefore with greater efficiency than the FFT or discrete Fourier transform (DFT). In the associated paper submitted to AAS, we use `PyStarshade` to assess the complementary role of a 60 m HWO starshade [@hwo_ss] paired with two different 6 m segmented and obscured telescope apertures featuring cm-scale details. We report $core$ $throughput$ for these apertures, defined as the fraction of exoplanet light recovered. Core throughput is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
+Fourier optics simulations of starshades and telescope apertures require fine spatial pixel scales to accurately model diffraction without aliasing errors. Traditional Fast Fourier Transforms (FFTs) necessitate large zero-padding factors to accomplish this, increasing computational cost. `PyStarshade` implements the Bluestein FFT [@bluestein] (related to the chirp Z-transform) to compute optical fields at arbitrary spatial scales, without zero-padding and therefore with greater efficiency than the FFT or discrete Fourier transform (DFT). In the associated paper submitted to AAS [@taaki_hwo_sim_2025], we use `PyStarshade` to assess the complementary role of a 60 m HWO starshade [@hwo_ss] paired with two different 6 m segmented and obscured telescope apertures featuring cm-scale details. We report $core$ $throughput$ for these apertures, defined as the fraction of exoplanet light recovered. Core throughput is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
 
 Furthermore, `PyStarshade` implements a novel technique to chunk Bluestein FFT calculations that mitigates memory bottlenecks. A set of unit tests are provided to validate the Fourier diffraction tools. `PyStarshade` pre-computes fields, point-spread-functions and throughput grids for different stages of the optical propagation chain, allowing for flexibility in modifying instrument parameters or telescope aperture masks. `PyStarshade` optionally interfaces with HCIPy [@hcipy] to generate telescope apertures and the resulting imaging. 
 
