+++ docs/paper/paper.md
@@ -42,7 +42,7 @@ Several starshade concepts have been proposed but not adopted, including a 26 m 
 # Statement of need
 `PyStarshade` provides a toolbox for performing optical simulations from source to focal plane with a configurable starshade telescope design. Complex electric fields are calculated at three planes of propagation (starshade, telescope aperture and focal plane) using Fresnel or Fraunhofer diffraction formula where appropriate. First-order imaging characteristics of a starshade can be determined from analytic relations which depend on the starshades size, telescope aperture size, wavelength and flight distance. `PyStarshade` allows for second-order imaging characteristics to be studied, including imaging simulations with a pixelized exoplanetary input scene of integrated flux per pixel and wavelength, varying starshade mask and telescope aperture mask to study throughput and post-processing methods. An example is provided in Figure 1, using spectral- and time-dependent  scenes consisting of planets, star and dust-disk, generated with `ExoVista` [@Stark_2022], the output imaging is shown in \autoref{fig:example}. `PyStarshade` is intended to be flexible and efficient in studying i) exoplanet retrievals, and, ii) instrument design.
 
-![Simulated imaging of a synthetic exoscene (ExoVista) with three visible exoplanets, as well as two planets within the region of starlight suppression, at a wavelength of 500 nm. The star to planet flux ratio of the visible planets range from $10^{-8}$ - $10^{-10}$. A 60 m HWO starshade configuration and a 6 m segmented pupil was used for this example. \label{fig:example}](exo_scene.png){ width=50% }
+![Simulated imaging of a synthetic exoscene (simulated with Exovista) with three visible exoplanets, as well as two planets within the region of starlight suppression, at a wavelength of 500 nm. The star to planet flux ratio of the visible planets range from $10^{-8}$ - $10^{-10}$. A 60 m HWO starshade configuration and a 6 m segmented pupil was used for this example. \label{fig:example}](exo_scene.png){ width=50% }
 
 Fourier optics simulations of starshades and telescope apertures require fine spatial pixel scales to accurately model diffraction without aliasing errors. Traditional Fast Fourier Transforms (FFTs) necessitate large zero-padding factors to accomplish this, increasing computational cost. `PyStarshade` implements the Bluestein FFT [@bluestein] (related to the chirp Z-transform) to compute optical fields at arbitrary spatial scales, without zero-padding and therefore with greater efficiency than the FFT or discrete Fourier transform (DFT). In the associated paper submitted to AAS [@taaki_hwo_sim_2025], we use `PyStarshade` to assess the complementary role of a 60 m HWO starshade [@hwo_ss] paired with two different 6 m segmented and obscured telescope apertures featuring cm-scale details. We report $core$ $throughput$ for these apertures, defined as the fraction of exoplanet light recovered. Core throughput is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
 
