+++ docs/paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-A key challenge in detecting exoplanets and characterizing their atmospheres is distinguishing an exoplanets faint reflected or emitted light from that of the star they orbit. Direct-imaging addresses this by suppressing light from the star to reveal orbiting exoplanet's [@direct]. Direct imaging can be achieved through two approaches: internal occulters, called coronagraphs, which use optics within the telescope to block starlight, and external occulters, called starshades, petal-shaped masks some 10's of meters across in diameter, flown in tandem with a telescope [@special]. Current direct imaging efforts using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, seeing a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel] in flux ratio between star and planet". 
+A key challenge in detecting exoplanets and characterizing their atmospheres is distinguishing an exoplanets faint reflected or emitted light from that of the star they orbit. Direct-imaging addresses this by suppressing light from the star to reveal orbiting exoplanet's [@direct]. Direct imaging can be achieved through two approaches: internal occulters, called coronagraphs, which use optics within the telescope to block starlight, and external occulters, called starshades, petal-shaped masks some 10's of meters across in diameter, flown in tandem with a telescope [@special]. Current direct imaging efforts using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, seeing a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{-8}$ - $10^{-10}$ [@angel] in flux ratio between star and planet. 
 
 While a starshade has not yet been flown in space, subscale starshades have demonstrated $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers like  water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
