### 2025-06-13 xiaziyna/PyStarshade#15 @xiaziyna
Hi, @ivalaginja - I think the issue may have been related to running out of memory causing Jupyter to grind to a halt when doing the optical simulations. This would be because all PSF's (of a given wavelength) are generated and held in memory while doing the optical calculations and written to disk at the end of the loop - I originally chose this because writing each term to disk can be slow itself (and also numpy npz compression doesn't support appending partially in this way). I have updated this to use hdf5, so it should never hold all the PSF's on memory and writes each one as it generates them.

Here's what I can't test/why it might break. At the end of generating the PSF basis, I then convert this to npz compressed format. However, to do this conversion, for a moment it has to hold the entire PSF basis in memory (which is large ~7GB). Your computer might not like this step - you'll know if it crashed here if the .h5 file was generated and is ~7GB, but no .npz file replaced it.

Note, I could change the spatial extent of the PSF to take up less memory, but I would prefer to keep things at a high level of accuracy (these samplings are based on attaining <0.01% error).

This takes again, just under 15 min on my computer. Just in case you happen have something running in the background using alot of memory: 'free -h	' to see how much is available.

Please try to run the updated version. It has a new requirement 'h5py' which is included in the setup.py:
'git lfs pull
pip install -e .'
