### 2025-06-14 xiaziyna/PyStarshade#15 @ivalaginja
@xiaziyna could it also be related to something else? I am now running the updated version and after 27 minutes of running, the `.h5` file has a size of a whopping 26.4 MB 😬 No wonder it never finished for me, at this rate it would take over 100 hours to get to a data size of 7 GB.

So it keeps updating indeed, but it is very, very slow... I have closed almost all of my other processes on my machine when running this. The python process for this is running at 100% CPU with ~11 GB of memory in the beginning, then dropping to 10 GB of memory.

I understand you want to keep a high fidelity for the PSF simulations, but is there anyway you can decrease that for the sake of having a notebook that is runnable for a random user that might not quite have a machine like you? You could always explain that the notebook runs a lower-fidelity case.

What are your thoughts?
