### 2025-06-14 #7917 @ivalaginja
@ivastar I might need help to compile the paper. The error message is: `failure' Downloading of the repository failed. Please make sure the URL is correct.`
