### 2025-06-14 xiaziyna/PyStarshade#15 @ivalaginja
> Hi [@ivalaginja](https://github.com/ivalaginja) - thanks for testing this for me. So it is having trouble doing the actual computations it seems. Can you do one other test for me if you get a chance and run this:
>
> import numpy as np, time, os
>
> N = 1000 a = (np.random.randn(N,N) + 1j*np.random.randn(N,N)).astype(np.complex128) start = time.time() _ = np.fft.fft2(a) print("wall-clock:", time.time()-start, "s")
>
> This will help me check if the FFT's are having trouble.

The result I get is `wall-clock: 0.038403987884521484 s`
