### 2025-06-15 xiaziyna/PyStarshade#15 @xiaziyna
Hi @ivalaginja - I believe I have identified/fixed the problem. The earlier cell where I construct the Exovista scene was using up and holding ~ 7 GB or more to generate the scene at some ~300 wavelengths. I have now freed this up (and just construct the scene at a single wavelength). Please try running the revised notebook. I will also consider changing the size of the PSF basis by reducing the precison, I just want to check that this is the issue before I go ahead and do that. Sorry about this!
