### 2025-06-16 xiaziyna/PyStarshade#15 @ivalaginja
Unfortunately I still see the same issue - after 8 hours and 20 minutes of runtime, using consistently about 2.3 GB of memory, the h5 file contains 371.4 MB of data, and hasn't finished.
