### 2025-06-16 xiaziyna/PyStarshade#15 @xiaziyna
I have been testing the code in Google Colab (with 12GB RAM). I think RAM is definitely not the issue and am also seeing about ~3GB of peak usage. After 30 min, about 1GB of data has been generated (which is expected as colab uses 1-2 CPU cores).

I wonder if the issue could be related to this (like maybe your FFT routines are only using a single CPU core)? Could you run another code snippet test for me:

  import numpy as np
  import os
  import psutil
  np.show_config()

  a = np.random.randn(2031, 2031)
  ft_a = np.fft.fft2(a)

  print("Threads used:", psutil.Process(os.getpid()).num_threads())

Thank you - I'm sorry about how annoying this has been!
