### 2025-06-16 xiaziyna/PyStarshade#15 @ivalaginja
@xiaziyna the output:
```
Build Dependencies:
  blas:
    detection method: pkgconfig
    found: true
    include directory: /private/var/folders/q3/_34lch1n3xsdf28hd1xbrxdh0000gn/T/cibw-run-v5l2ycuc/cp313-macosx_x86_64/build/venv/lib/python3.13/site-packages/scipy_openblas64/include
    lib directory: /private/var/folders/q3/_34lch1n3xsdf28hd1xbrxdh0000gn/T/cibw-run-v5l2ycuc/cp313-macosx_x86_64/build/venv/lib/python3.13/site-packages/scipy_openblas64/lib
    name: scipy-openblas
    openblas configuration: OpenBLAS 0.3.28  USE64BITINT DYNAMIC_ARCH NO_AFFINITY
      Sandybridge MAX_THREADS=64
    pc file directory: /Users/runner/work/numpy/numpy/.openblas
    version: 0.3.28
  lapack:
    detection method: pkgconfig
    found: true
    include directory: /private/var/folders/q3/_34lch1n3xsdf28hd1xbrxdh0000gn/T/cibw-run-v5l2ycuc/cp313-macosx_x86_64/build/venv/lib/python3.13/site-packages/scipy_openblas64/include
    lib directory: /private/var/folders/q3/_34lch1n3xsdf28hd1xbrxdh0000gn/T/cibw-run-v5l2ycuc/cp313-macosx_x86_64/build/venv/lib/python3.13/site-packages/scipy_openblas64/lib
    name: scipy-openblas
    openblas configuration: OpenBLAS 0.3.28  USE64BITINT DYNAMIC_ARCH NO_AFFINITY
      Sandybridge MAX_THREADS=64
    pc file directory: /Users/runner/work/numpy/numpy/.openblas
    version: 0.3.28
Compilers:
  c:
    commands: cc
    linker: ld64
    name: clang
    version: 15.0.0
  c++:
    commands: c++
    linker: ld64
    name: clang
    version: 15.0.0
  cython:
    commands: cython
    linker: cython
    name: cython
    version: 3.0.12
Machine Information:
  build:
    cpu: x86_64
    endian: little
    family: x86_64
    system: darwin
  host:
    cpu: x86_64
    endian: little
    family: x86_64
    system: darwin
Python Information:
  path: /private/var/folders/q3/_34lch1n3xsdf28hd1xbrxdh0000gn/T/build-env-cwgn10z4/bin/python
  version: '3.13'
SIMD Extensions:
  baseline:
  - SSE
  - SSE2
  - SSE3
  found:
  - SSSE3
  - SSE41
  - POPCNT
  - SSE42
  not found:
  - AVX
  - F16C
  - FMA3
  - AVX2
  - AVX512F
  - AVX512CD
  - AVX512_KNL
  - AVX512_SKX
  - AVX512_CLX
  - AVX512_CNL
  - AVX512_ICL

Threads used: 10
```
