### 2025-06-17 xiaziyna/PyStarshade#15 @xiaziyna
I appreciate you taking the time to run these tests and respond quickly!

I also want to understand why this isn't running on your machine.

So it's not a RAM issue, and a single 1000x1000 FFT runs fine (your time test averaged at about 0.03s, mine is also about this speed). The loop in the notebook essentially computes many FFTs (each of a smaller size than the test ~ 700 x 700).

Your M1 chip should be quite efficient (it has 8 cores, the test indicated 10 threads were used, for comparison I have 12 cores and 16 threads were used and Intel is supposed to be worse). It's strange that after 30 min, 100 MB were generated, but after 8 hours, only ~ 400 MB were generated - this indicates the computations aren't just ultra slow but they're actively slowing down/hitting a wall. This shouldn't happen as our CPU architectures are quite comparable.

I ran the same code in colab and it took about 7000s (bottom right corner) or <2 hr to complete (colab has 1-2 cores available).

![Image](https://github.com/user-attachments/assets/8749a7fb-1ff4-47ac-98f9-18f2cbc9da4c)

https://colab.research.google.com/drive/1G8zkb2Hz-oT-ZvEXv-P8RNSzOamo7NvW?usp=sharing

The only thing I can think of is that maybe the management of threads is poor as more and more FFT's are executed. One thing I notice is your blas build is compiled for sandybridge which is an intel chip - so this might be a problem.
Here is another test to check if doing a large number of FFT's is 'clogging' your system:

import numpy as np
import time

size = 1001
batch_sizes = [1, 10, 100, 1000, 10000]

for batch in batch_sizes:
    start = time.time()
    for i in range(batch):
        data = np.random.random((size, size))
        np.fft.fft2(data)
    end = time.time()

    time_per_fft = (end - start) / batch
    print(f"Batch {batch}: {time_per_fft:.4f} seconds per FFT")

For reference, my output was:
Batch 1: 0.0291 seconds per FFT
Batch 10: 0.0275 seconds per FFT
Batch 100: 0.0269 seconds per FFT
Batch 1000: 0.0270 seconds per FFT
Batch 10000: 0.0329 seconds per FFT

Just to check - have you by any chance ran out of hard-drive space?
