+++ docs/paper/paper.bib
@@ -170,7 +170,7 @@ URL = {https://doi.org/10.1117/1.JATIS.7.2.021211}
 @software{fresnaq,
   author  = {Barnett, Alex H.},
   title   = {fresnaq: fast Fresnel scalar diffraction from binary occulters and apertures},
-  version = {5181d86},              % commit hash
+  version = {5181d86},
   date    = {2021-01-13},
   url     = {https://github.com/ahbarnett/fresnaq},
   note    = {GitHub repository, commit 5181d86; accessed 2025-06-13}
@@ -227,7 +227,7 @@ URL = {https://doi.org/10.1117/1.JATIS.7.2.021211}
 @software{diffraq,
   author  = {Harness, Anthony},
   title   = {diffraq: Diffraction with Areal Quadrature},
-  version = {a61613a},            % commit hash (Jul 26 2021)
+  version = {a61613a},
   year    = {2021},
   url     = {https://github.com/harnessa/diffraq/tree/a61613a},
   note    = {GitHub repository, commit a61613a; accessed 2025-06-13}
