### 2025-06-17 xiaziyna/PyStarshade#19 @ivalaginja
This is great. A last couple of typos that I found, but the only real weird sentence is in line 47 (see below). If you could fix at least that, we can call this issue done.

9. "an exoplanet's" needs an apostrophe
23. add comma after "suppression"
31. no comma here
46. missing "a" in front of "dust disk"
47. Can delete: "the output imaging is shown in Figure 1." and link to the figure in the "Figure 1" in line 45 instead
Fig. 1: I think you are using "starshade" without a hyphen everywhere else?
Fig. 1: ->"between 10^8 and 10^10." ("and" should be in there)
