### 2025-06-17 xiaziyna/PyStarshade#19 @xiaziyna
Thanks - I have updated the paper with these changes.
