### 2025-06-17 xiaziyna/PyStarshade#15 @ivalaginja
> For reference, my output was:
Batch 1: 0.0291 seconds per FFT
Batch 10: 0.0275 seconds per FFT
Batch 100: 0.0269 seconds per FFT
Batch 1000: 0.0270 seconds per FFT
Batch 10000: 0.0329 seconds per FFT

My results are comparable so that indeed doesn't seem to be it:

Batch 1: 0.1903 seconds per FFT
Batch 10: 0.0248 seconds per FFT
Batch 100: 0.0229 seconds per FFT
Batch 1000: 0.0233 seconds per FFT
Batch 10000: 0.0235 seconds per FFT
