### 2025-06-18 xiaziyna/PyStarshade#15 @ivalaginja
> Hi [@ivalaginja](https://github.com/ivalaginja)
>
> I think I have another guess at the problem - I had removed all instances of float128 type before when it caused issued for [@Jashcraf](https://github.com/Jashcraf) (this type is not supported on Mac). One of the data-files used here is complex256 type - I think this might be the issue.
>
> Quick test for this:
>
> ```
> import numpy as np
> import time
>
> N = 3051
> a = (np.random.randn(N,N) + 1j*np.random.randn(N,N)).astype(np.complex256)
> start = time.time()
> b = np.fft.fft2(a)
> print("wall-clock:", time.time()-start, "s")
> ```
>
> 0.7s on my machine - if it runs extraordinarily slower or crashes then this indicates the type might be a problem (I will convert the data to complex128 in any case).

Lol `wall-time: 81.38478391699937 s`
