### 2025-06-18 xiaziyna/PyStarshade#15 @Jashcraf
Hopping on this now - tried running the example notebook as-is. Had the following requests:

### 1) There appears to be an import error at the second cell
```python
import download_exovista
```

This is because the example notebook isn't in the same folder as `download_exovista.py`, which is an example file. It'd be preferable if this was imported from the Pystarshade source if it's going to be used frequently. However, a quick fix is
```python
from examples import download_exovista
```

Which worked for me

### 2) Encountered the np.float128 error discussed above in StarshadeProp

The following throws an error for np.float128 which @xiaziyna mentioned doesn't work on M1 macs at least. Does that mean all instances of np.float128 haven't been eliminated? I'll try run on my linux machine when I get a chance

```python
drm = 'wfirst'
wfirst_starshade = StarshadeProp(drm = drm)
```
