### 2025-06-18 xiaziyna/PyStarshade#15 @xiaziyna
Hi @Jashcraf

1) A copy of the notebook is kept in the examples folder - I also put a copy in the folder for the documentation so that it could render (I can modify this version to import from the correct place).

2) I wrote a script before to find all instances/data that used float128 but did not do the same for complex256 - I am currently pushing a corrected version but had some trouble with git.
