+++ docs/index.rst
@@ -33,16 +33,16 @@ Table of Contents
    :maxdepth: 2
 
    PyStarshade Documentation Landing Page <https://pystarshade.readthedocs.io>
-   content/install       
-   content/background   
-   content/usage        
+   docs/content/install       
+   docs/content/background   
+   docs/content/usage        
 
 .. toctree::
    :maxdepth: 1
 
-    ../examples/solar_system
-   content/test
-   content/contribute
+   examples/solar_system
+   docs/content/test
+   docs/content/contribute
 
 
 
