+++ docs/content/solar_system_example.rst
@@ -1,0 +1,4 @@
+Solar System Example
+====================
+
+.. nbinclude:: ../../examples/solar_system.ipynb
