+++ docs/conf.py
@@ -17,8 +17,9 @@ release = '1.0'
 # paths from conf.py
 import sys
 import os
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../examples')))
+
 
 # Add the examples directory to the path
 #sys.path.insert(0, os.path.abspath('../examples'))
@@ -36,10 +37,6 @@ extensions = [    'sphinx.ext.autodoc',	     # To generate autodocs
     'sphinx.ext.napoleon',           # For auto-doc configuration
 ]
 
-# Configure nbsphinx to find notebooks elsewhere
-nbsphinx_custom_formats = {
-    '.ipynb': ['jupyter', 'notebook']
-}
 
 napoleon_google_docstring = False   # Turn off googledoc strings
 napoleon_numpy_docstring = True     # Turn on numpydoc strings
