### 2025-06-18 xiaziyna/PyStarshade#15 @xiaziyna
Also @Jashcraf, if you run the following I can see if there's any weird setting to only use a single thread:

import numpy as np
np.show_config()

Thanks!
