+++ docs/.readthedocs.yaml
@@ -1,0 +1,14 @@
+version: 2
+
+sphinx:
+  configuration: docs/conf.py
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+
+python:
+  install:
+    - method: pip
+      requirements: docs/requirements.txt
