+++ docs/conf.py
@@ -44,7 +44,6 @@ napoleon_use_ivar = True 	     # For maths symbology
 imgmath_image_format = 'svg'
 highlight_language = 'python'
 
-nbsphinx_allow_errors = True
 
 #imgmath_latex_preamble = r"""
 #\usepackage{amsmath}
