### 2025-06-18 xiaziyna/PyStarshade#15 @ivalaginja
@xiaziyna I can probably fix this with a local fix, but is this behavior intended?

```
Cell In[10], line 6
      4 # generating the PSF basis at a single wavelength
      5 hwo_starshade.N_wl = 1 # edit an internal variable so that only a single wavelength PSF basis is calculated
----> 6 hwo_starshade.gen_psf_basis(pupil_type = pupil_type)
      7 print ('psf generation complete!')
      9 end = time.time()

File ~/repos/PyStarshade/pystarshade/propagator.py:232, in StarshadeProp.gen_psf_basis(self, pupil_type, pupil_symmetry)
    230 psf_points = flat_grid(N_basis, negative = 1 - pupil_symmetry)
    231 for wl_i in range(self.N_wl):
--> 232     data = np.load(data_file_path(self.drm+'_pupil'+'_'+self.d_x_str+'_'+str(int(self.wl_range[wl_i]*1e9))+'.npz', 'fields'))
    233     wl = data['params'][0]
    234     save_path_h5 = data_file_path(self.drm+'_psf_'+pupil_type+'_'+self.d_x_str+'_'+str(int(wl * 1e9))+'.h5','psf')

File ~/opt/miniconda3/envs/pystarshade/lib/python3.13/site-packages/numpy/lib/_npyio_impl.py:486, in load(file, mmap_mode, allow_pickle, fix_imports, encoding, max_header_size)
    483 else:
    484     # Try a pickle
    485     if not allow_pickle:
--> 486         raise ValueError(
    487             "This file contains pickled (object) data. If you trust "
    488             "the file you can load it unsafely using the "
    489             "`allow_pickle=` keyword argument or `pickle.load()`.")
    490     try:
    491         return pickle.load(fid, **pickle_kwargs)

ValueError: This file contains pickled (object) data. If you trust the file you can load it unsafely using the `allow_pickle=` keyword argument or `pickle.load()`.
```
