### 2025-06-18 xiaziyna/PyStarshade#15 @xiaziyna
You're going to get an error from float128 params leftover in this file - I need to update them again, so no need to bother updating/testing for now. I will have a working version done for tomorrow.
