+++ docs/requirements.txt
@@ -5,3 +5,4 @@ nbsphinx
 ipykernel
 scipy
 hcipy
+h5py
