+++ docs/index.rst
@@ -42,6 +42,7 @@ Table of Contents
 .. toctree::
    :maxdepth: 1
 
+   content/solar_system_example
    content/test
    content/contribute
 
