+++ docs/content/solar_system_example.rst
@@ -1,8 +1,0 @@
-Solar System Example
-====================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Examples:
-
-   ../examples/solar_system
