+++ docs/index.rst
@@ -25,8 +25,6 @@ If you use ``PyStarshade`` in your research, please include the following citati
 
     Taaki, Kamalabadi, Kemball. *PyStarshade: Simulating High-Contrast Imaging of Exoplanets with Starshades*.
 
-.. nbinclude:: ../../examples/solar_system.ipynb
-
 
 Table of Contents
 -----------------
