### 2025-06-19 xiaziyna/PyStarshade#15 @ivalaginja
@xiaziyna just to check, did you bifurcate your branches on purpose? i.e., the v1.0.0 tag is not in the history of the main branch anymore?
