### 2025-06-19 xiaziyna/PyStarshade#15 @ivalaginja
And the GMT cell takes me 14-15 sec to run. I am impressed by how you churned through this!
