+++ docs/paper/paper.md
@@ -33,23 +33,23 @@ bibliography: paper.bib
 
 # Summary
 
-A key challenge in detecting exoplanets and characterizing their atmospheres is distinguishing an exoplanet's faint reflected or emitted light from that of the star they orbit. Direct-imaging addresses this by suppressing light from the star to reveal orbiting exoplanet's [@direct]. Direct imaging can be achieved through two approaches: internal occulters, called coronagraphs, which use optics within the telescope to block starlight, and external occulters, called starshades, petal-shaped masks some 10's of meters across in diameter, flown in tandem with a telescope [@special]. Current direct imaging efforts using ground-based coronagraphs like the Very Large Telescope (VLT) and Gemini Observatory have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, seeing a faint Earth-like planet in the habitable zone of a Sun-like star will require extreme starlight suppression to the level of $10^{8}$ - $10^{10}$ star-to-planet flux ratio [@angel]. 
+A key challenge in detecting exoplanets and characterizing their atmospheres is distinguishing their faint reflected or emitted light from that of their host star. Direct imaging addresses this challenge by suppressing the star's light to reveal the orbiting exoplanet. There are two approaches to direct imaging: internal occulters, called coronagraphs, which use optics within the telescope to block starlight; and external occulters, called starshades, which are petal-shaped masks spanning tens of meters in diameter, flown in tandem with a telescope [@special]. Current direct imaging efforts using ground-based coronagraphs at the Very Large Telescope (VLT) and at the Gemini Observatory have successfully imaged young, massive exoplanets at wide separations [@Follette_2023]. However, observing a faint, Earth-like planet in the habitable zone of a Sun-like star requires extreme starlight suppression to the level of $10^{8}$ - $10^{10}$ star-to-planet flux ratio [@angel].
 
-While a starshade has not yet been flown in space, subscale starshades have demonstrated $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers like  water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression, critical questions remain regarding the properties and orbital configurations of exoplanetary systems observable by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox to efficiently perform optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
+Although no starshade has been flown in space yet, subscale starshades have demonstrated $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers such as water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression, critical questions remain regarding the properties and orbital configurations of exoplanetary systems that can be observed by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox that efficiently performs optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
[…]
