### 2025-07-02 86cce56
Update paper.bib
