+++ docs/paper/paper.md
@@ -49,7 +49,7 @@ Fourier optics simulations of starshades and telescope apertures require fine sp
 Furthermore, `PyStarshade` implements a novel technique to chunk Bluestein FFT calculations, which mitigates memory bottlenecks. A set of unit tests is provided to validate the Fourier diffraction tools. `PyStarshade` precomputes fields, point spread functions, and throughput grids for different stages of the optical propagation chain. This allows flexibility in modifying instrument parameters or telescope aperture masks. `PyStarshade` optionally interfaces with HCIPy [@hcipy] to generate telescope apertures and the resulting imaging.
 
 # Related software
-The Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance (`SISTER`) [@sisters:2022], implemented in MATLAB, performs detailed, end-to-end simulations of starshade imaging, further including instrument noise sources such as solar glint, and tools for modeling an exoplanetary system and background objects. The optical simulation pipeline of `PyStarshade` follows a similar approach to that of `SISTER`, employing a precomputed PSF basis for image simulation. However, SISTER computes diffraction from the starshade to the telescope aperture using the boundary diffraction wave method [@cady], performing a 1D integral along the starshade's edge, followed by a DFT to obtain the final PSF basis. Several other libraries focus exclusively on calculating the starshade diffraction incident on the telescope aperture. One such library is `fresnaq`, a MATLAB library that efficiently computes diffraction by a starshade via the non-uniform FFT [@barnett]. `diffraq` [@diffraq] is a Python implementation of `freqnaq`. This library also provides an implementation of the boundary diffraction wave method and tools to generate starshade masks [@harness]. The starshade masks used in `PyStarshade`/ are obtained from `SISTER` or were generated with `diffraq`. Outside of imaging simulations, the Exoplanet Open-Source Imaging Mission Simulator (`EXOSIMS`) [@exosims] provides methods for orbital simulation of starshades and fuel consumption calculations.
+The Starshade Imaging Simulation Toolkit for Exoplanet Reconnaissance (`SISTER`) [@sisters:2022], implemented in MATLAB, performs detailed, end-to-end simulations of starshade imaging, further including instrument noise sources such as solar glint, and tools for modeling an exoplanetary system and background objects. The optical simulation pipeline of `PyStarshade` follows a similar approach to that of `SISTER`, employing a precomputed PSF basis for image simulation. However, SISTER computes diffraction from the starshade to the telescope aperture using the boundary diffraction wave method [@cady], performing a 1D integral along the starshade's edge, followed by a DFT to obtain the final PSF basis. Several other libraries focus exclusively on calculating the starshade diffraction incident on the telescope aperture. One such library is `fresnaq`, a MATLAB library that efficiently computes diffraction by a starshade via the non-uniform FFT [@barnett]. `diffraq` [@diffraq] is a Python implementation of `freqnaq`. This library also provides an implementation of the boundary diffraction wave method and tools to generate starshade masks [@harness]. The starshade masks used in `PyStarshade` are obtained from `SISTER` or were generated with `diffraq`. Outside of imaging simulations, the Exoplanet Open-Source Imaging Mission Simulator (`EXOSIMS`) [@exosims] provides methods for orbital simulation of starshades and fuel consumption calculations.
 
 # Acknowledgements
 
