### 2025-07-02 e3e1262
Update journal names
