+++ docs/paper/paper.bib
@@ -97,7 +97,7 @@ abstract = {This tutorial is an introduction to High-Contrast Imaging, a techniq
 @ARTICLE{angel,
        author = {{Angel}, J.~R.~P. and {Cheng}, A.~Y.~S. and {Woolf}, N.~J.},
         title = "{A space telescope for infrared spectroscopy of Earth-like planets}",
-      journal = {\nat},
+      journal = {Nature},
      keywords = {Astronomical Spectroscopy, Extrasolar Planets, Infrared Astronomy, Spaceborne Telescopes, Apertures, Diffraction Patterns, Light (Visible Radiation), Planetology, Radial Distribution, Astronomy},
          year = 1986,
         month = jul,
