+++ docs/paper/paper.bib
@@ -239,6 +239,29 @@ journal = {The Astronomical Journal},
 abstract = {Studies of future space- and ground-based exoplanet surveys often rely on models of planetary systems to simulate instrument response, estimate scientific yields, perform trade analyses, and study efficient observation strategies. Until now, no planetary system models contained all of the basic physics necessary to enable study with all of the major exoplanet detection methods. Here we introduce a suite of such models generated by a new tool, exoVista. The exoVista tool quickly generates thousands of models of quasi-self-consistent planetary systems around known nearby stars at scattered light wavelengths and efficiently records the position, velocity, spectrum, and physical parameters of all bodies as functions of time. The modeled planetary systems can be used to simulate surveys using the direct imaging, transit, astrometric, and radial velocity techniques, as well as the overlap of these different methods.}
 }
 
+@Article{         harris2020array,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
 @software{exovista,
   author   = {Stark, Christopher C. and Howe, Alex},
   title    = {ExoVista: Planetary System Models for Exoplanet Studies},
