+++ docs/paper/paper.md
@@ -37,7 +37,7 @@ A key challenge in detecting exoplanets and characterizing their atmospheres is 
 
 Although no starshade has been flown in space yet, subscale starshades have demonstrated $\leq 10^{9}$ starlight suppression at a wavelength of 633 nm in laboratory demonstrations [@subscale], and $\leq 10^{10}$ across the broad wavelength range necessary for detecting biomarkers such as water, oxygen, and methane in computational simulations [@mennesson]. Beyond demonstrating starlight suppression, critical questions remain regarding the properties and orbital configurations of exoplanetary systems that can be observed by starshade missions, as well as optimal instrument designs. `PyStarshade` is a Python toolbox that efficiently performs optical simulations of starshades with exoplanetary scenes to assess the utility of starshades in future direct-imaging missions.
 
-Several starshade concepts have been proposed but not yet adopted. These include: a 26-meter-diameter design to rendezvous with the Nancy Grace Roman Space Telescope (scheduled for launch in 2027) [@ngrst_rendezvous]; a 52-meter design as part of the Habex concept [@habex];  a 60-meter starshade operating over visible-to-infrared wavelengths 500–1,000 nm) and a 35-meter UV (250–500 nm) starshade, proposed for the Habitable Worlds Observatory (HWO), a NASA mission is in the early design phase, aiming to find and characterize a handful of Earth-like exoplanets [@NAP26141]. An example use of  `PyStarshade` to evaluate the complementary role of a 60-meter HWO starshade [@hwo_ss] paired with two different 6-meter segmented and obscured telescope apertures featuring centimeter-scale details is presented in [@taaki_hwo_sim_2025] where we report $core$ $throughput$ for these apertures. Core throughput is defined as the fraction of exoplanet light recovered and is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
+Several starshade concepts have been proposed but not yet adopted. These include: a 26-meter-diameter design to rendezvous with the Nancy Grace Roman Space Telescope (scheduled for launch in 2026) [@ngrst_rendezvous]; a 52-meter design as part of the Habex concept [@habex];  a 60-meter starshade operating over visible-to-infrared wavelengths 500–1,000 nm) and a 35-meter UV (250–500 nm) starshade, proposed for the Habitable Worlds Observatory (HWO), a NASA mission is in the early design phase, aiming to find and characterize a handful of Earth-like exoplanets [@NAP26141]. An example use of  `PyStarshade` to evaluate the complementary role of a 60-meter HWO starshade [@hwo_ss] paired with two different 6-meter segmented and obscured telescope apertures featuring centimeter-scale details is presented in [@taaki_hwo_sim_2025] where we report $core$ $throughput$ for these apertures. Core throughput is defined as the fraction of exoplanet light recovered and is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
 
 # Statement of need
[…]
