+++ docs/paper/paper.md
@@ -40,7 +40,7 @@ Although no starshade has been flown in space yet, subscale starshades have demo
 Several starshade concepts have been proposed but not yet adopted. These include: a 26-meter-diameter design to rendezvous with the Nancy Grace Roman Space Telescope (scheduled for launch in 2026) [@ngrst_rendezvous]; a 52-meter design as part of the Habex concept [@habex];  a 60-meter starshade operating over visible-to-infrared wavelengths 500–1,000 nm) and a 35-meter UV (250–500 nm) starshade, proposed for the Habitable Worlds Observatory (HWO), a NASA mission is in the early design phase, aiming to find and characterize a handful of Earth-like exoplanets [@NAP26141]. An example use of  `PyStarshade` to evaluate the complementary role of a 60-meter HWO starshade [@hwo_ss] paired with two different 6-meter segmented and obscured telescope apertures featuring centimeter-scale details is presented in [@taaki_hwo_sim_2025] where we report $core$ $throughput$ for these apertures. Core throughput is defined as the fraction of exoplanet light recovered and is a key metric that governs the exposure time needed to image an exoplanet and therefore the overall exoplanetary yield of a mission.
 
 # Statement of need
-`PyStarshade` provides a toolbox for performing optical simulations from source to focal plane with a configurable starshade telescope design. Complex electric fields are calculated at three planes of propagation: the starshade, the telescope aperture, and the focal plane, using the Fresnel or Fraunhofer diffraction formula where appropriate. First-order  imaging characteristics of a starshade can be determined from analytic relations that depend on the size of the starshade, the size of the telescope aperture, the wavelength of the observations, and the flight distance. `PyStarshade` allows for second-order imaging characteristics to be studied, including imaging simulations with a pixelized exoplanetary input scene of integrated flux per pixel and wavelength. It also allows for the study of throughput and post-processing methods with varying starshade masks and telescope aperture masks to study throughput and post-processing methods. An example imaged scene at a snapshot in time and wavelength is shown in \autoref{fig:example},/Fig. 1, consisting of a planet, a star, and a dust disk. These spectral- and time-dependent scenes were generated with ExoVista [@Stark_2022]. `PyStarshade` is intended to be flexible and efficient in studying exoplanet retrievals and instrument design.
+`PyStarshade` provides a toolbox for performing optical simulations from source to focal plane with a configurable starshade telescope design. Complex electric fields are calculated at three planes of propagation: the starshade, the telescope aperture, and the focal plane, using the Fresnel or Fraunhofer diffraction formula where appropriate. First-order  imaging characteristics of a starshade can be determined from analytic relations that depend on the size of the starshade, the size of the telescope aperture, the wavelength of the observations, and the flight distance. `PyStarshade` allows for second-order imaging characteristics to be studied, including imaging simulations with a pixelized exoplanetary input scene of integrated flux per pixel and wavelength. It also allows for the study of throughput and post-processing methods with varying starshade masks and telescope aperture masks to study throughput and post-processing methods. An example imaged scene is shown in \autoref{fig:example},/Fig. 1 at a snapshot in time and wavelength. The scene consists of a planet, a star, and a dust disk. These spectral- and time-dependent scenes were generated with ExoVista [@Stark_2022]. `PyStarshade` is intended to be flexible and efficient in studying exoplanet retrievals and instrument design.
 
 ![A starshade imaging simulation shown at a wavelength of 500 nm with a synthetic exoplanetary input scene (generated with ExoVista): three exoplanets are directly visible, while two more sit inside the starshade suppression zone. The scene assumes a 60 m HWO starshade paired with a 6 m segmented telescope; the planets in the scene have planet-to-star flux ratios between $10^{-8}$ and $10^{-10}$. \label{fig:example}](exo_scene.png){ width=50% }
 
