+++ docs/paper/paper.bib
@@ -239,7 +239,7 @@ journal = {The Astronomical Journal},
 abstract = {Studies of future space- and ground-based exoplanet surveys often rely on models of planetary systems to simulate instrument response, estimate scientific yields, perform trade analyses, and study efficient observation strategies. Until now, no planetary system models contained all of the basic physics necessary to enable study with all of the major exoplanet detection methods. Here we introduce a suite of such models generated by a new tool, exoVista. The exoVista tool quickly generates thousands of models of quasi-self-consistent planetary systems around known nearby stars at scattered light wavelengths and efficiently records the position, velocity, spectrum, and physical parameters of all bodies as functions of time. The modeled planetary systems can be used to simulate surveys using the direct imaging, transit, astrometric, and radial velocity techniques, as well as the overlap of these different methods.}
 }
 
-@Article{         harris2020array,
+@Article{         numpy,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
                  van der Walt and Ralf Gommers and Pauli Virtanen and David
