### 2025-07-09 fd48ed9
paper edits
