+++ docs/paper/paper.bib
@@ -13,6 +13,50 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{scipy,
+  title={{SciPy} 1.0: Fundamental algorithms for scientific computing in Python},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature Methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
+}
+
+@article{astropy,
+  title={Astropy: A community Python package for astronomy},
+  author={Astropy Collaboration and Robitaille, Thomas P. and Tollerud, Erik J. and Greenfield, Perry and Droettboom, Michael and Bray, Erik and Aldcroft, Tom and Davis, Matt and Ginsburg, Adam and Price-Whelan, Adrian M. and others},
+  journal={Astronomy \& Astrophysics},
+  volume={558},
+  pages={A33},
+  year={2013},
+  doi={10.1051/0004-6361/201322068}
+}
+
+@misc{setuptools,
+  title={{setuptools}: Easily download, build, install, upgrade, and uninstall Python packages},
+  author={{Python Packaging Authority}},
+  year={2024},
+  note={\url{https://github.com/pypa/setuptools}}
+}
+
+@misc{pytest,
+  title={pytest: simple powerful testing with Python},
+  author={Krekel, Holger and others},
+  year={2024},
+  note={\url{https://docs.pytest.org}}
+}
+
+@misc{h5py,
+  title={{h5py}: HDF5 for Python},
+  author={Collette, Andrew and contributors},
+  year={2024},
+  note={\url{https://www.h5py.org}}
+}
+
+
 
 @INCOLLECTION{direct,
        author = {{Traub}, W.~A. and {Oppenheimer}, B.~R.},
