### 2025-07-09 affb7dc
paper edits
