### 2025-07-09 #7917 @xiaziyna
Hi @ivastar,

I have completed the requested changes and released a version/archived this on Zenodo. The author lists are the same and the licenses match.

[10.5281/zenodo.15849727](https://doi.org/10.5281/zenodo.15849727)
