### 2025-07-11 #7917 @ivastar
> * Make sure that the **title** and author list (including ORCIDs) in the archive match those in the JOSS paper.

@xiaziyna can you please make sure that the Zenodo title also matches the title of the paper?
