### 2025-07-16 #7917 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
