### 2025-07-16 #7917 @warrickball
Hi @xiaziyna, I'm the associate EiC for this subject area and will do some editorial checks before we publish the paper. I'll try to get to those in the next few working days.

In the meantime, our contacts at AAS Journals have brought the joint submission back to our attention and there's been a longer discussion between the JOSS and AAS Journal editorial teams. We *can* still associate our AAS Journals submission with the JOSS submission if you like. The only real requirement will be providing AAS Journals with the JOSS DOI. My current understanding is that we can add the AAS Journals DOI to the paper even after publication in JOSS (but I'll make sure this is crystal clear before doing anything that would disrupt the association between the articles).

I want to be clear that this differs from @ivastar's advice during the review only because the guidance we'd given to our JOSS editors was unclear. This update comes from discussions that @ivastar wasn't party to. We'll be updating our documentation to make this clearer for our editors.
