### 2025-07-16 #7917 @xiaziyna
Hi @warrickball

Thank you for taking over for the editorial checks!

I would like to associate the JOSS and AAS submissions if possible. Would you need anything on my end?

I would especially like to thank the reviewers and editor for their work on this - which I recognize has been substantial. Thank you @ivastar for overseeing the review and for your careful editorial work on the paper. Thank you @ivalaginja and @Jashcraf for your sharing your expertise and feedback, which have improved the overall utility of the tool!
