### 2025-07-20 #7917 @warrickball
> I would like to associate the JOSS and AAS submissions if possible. Would you need anything on my end?

Not at this point. When the AAS publication has a DOI, we'll add it to the JOSS paper but this can happen even after we've accepted in the paper in JOSS, since all we do is update some metadata. One of the AAS Journals data editors, @augustfly, is already in the loop.
