### 2025-07-20 #7917 @warrickball
I'll probably open a PR with some editorial tweaks later this week but I already noticed that you don't have an affiliation or ORCID in the Zenodo archive. You can update the metadata without making a new submission.
