### 2025-07-23 c61e893
Initial editorial tweaks for JOSS publication
