+++ docs/paper/paper.bib
@@ -14,7 +14,7 @@
 }
 
 @article{scipy,
-  title={{SciPy} 1.0: Fundamental algorithms for scientific computing in Python},
+  title={{SciPy} 1.0: Fundamental algorithms for scientific computing in {Python}},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature Methods},
   volume={17},
@@ -26,7 +26,7 @@
 }
 
 @article{astropy,
-  title={Astropy: A community Python package for astronomy},
+  title={Astropy: A community {Python} package for astronomy},
   author={Astropy Collaboration and Robitaille, Thomas P. and Tollerud, Erik J. and Greenfield, Perry and Droettboom, Michael and Bray, Erik and Aldcroft, Tom and Davis, Matt and Ginsburg, Adam and Price-Whelan, Adrian M. and others},
   journal={Astronomy \& Astrophysics},
   volume={558},
@@ -43,20 +43,20 @@
 }
 
 @misc{pytest,
-  title={pytest: simple powerful testing with Python},
-  author={Krekel, Holger and others},
-  year={2024},
-  note={\url{https://docs.pytest.org}}
+  author       = {Holger Krekel and Bruno Oliveira and Ronny Pfannschmidt and Floris Bruynooghe and Brianna Laugher and Florian Bruhin},
+  title        = {pytest x.y},
+  year         = {2004},
+  howpublished = {\url{https://github.com/pytest-dev/pytest}},
+  note         = {Version x.y. Contributors include Holger Krekel, Bruno Oliveira, Ronny Pfannschmidt, Floris Bruynooghe, Brianna Laugher, Florian Bruhin, and others.}
 }
 
-@misc{h5py,
-  title={{h5py}: HDF5 for Python},
-  author={Collette, Andrew and contributors},
-  year={2024},
-  note={\url{https://www.h5py.org}}
+@Book{h5py,
+    keywords = {python, hdf5},
+        year = {2013},
+   publisher = {O'Reilly},
+       title = {Python and HDF5},
+      author = {Andrew Collette}
 }
-
-
 
 @INCOLLECTION{direct,
        author = {{Traub}, W.~A. and {Oppenheimer}, B.~R.},
@@ -87,7 +87,7 @@ URL = {https://doi.org/10.1117/1.JATIS.7.2.021207}
 
 @article{special,
 author = {Jonathan W. Arenberg and Anthony D. Harness and Rebecca M. Jensen-Clem},
-title = {{Special Section on Starshades: Overview and a Dialogue}},
+title = {Special Section on Starshades: Overview and a Dialogue},
 volume = {7},
 journal = {Journal of Astronomical Telescopes, Instruments, and Systems},
 number = {2},
@@ -128,8 +128,8 @@ abstract = {This tutorial is an introduction to High-Contrast Imaging, a techniq
 
 @ARTICLE{bluestein,
   author={Bluestein, L.},
-  journal={IEEE Transactions on Audio and Electroacoustics}, 
-  title={A linear filtering approach to the computation of discrete Fourier transform}, 
+  journal={IEEE Transactions on Audio and Electroacoustics},
+  title={A linear filtering approach to the computation of discrete {Fourier} transform},
   year={1970},
   volume={18},
   number={4},
@@ -171,7 +171,7 @@ URL = {https://doi.org/10.1117/1.JATIS.10.3.035004}
 
 @inproceedings{hwo_ss,
[…]
