+++ docs/paper/paper.bib
@@ -44,7 +44,7 @@
 
 @misc{pytest,
   author       = {Holger Krekel and Bruno Oliveira and Ronny Pfannschmidt and Floris Bruynooghe and Brianna Laugher and Florian Bruhin},
-  title        = {pytest x.y},
+  title        = {pytest 7.4},
   year         = {2004},
   howpublished = {\url{https://github.com/pytest-dev/pytest}},
   note         = {Version x.y. Contributors include Holger Krekel, Bruno Oliveira, Ronny Pfannschmidt, Floris Bruynooghe, Brianna Laugher, Florian Bruhin, and others.}
