### 2025-07-25 #7917 @warrickball
Thanks once again to @Jashcraf & @ivalaginja for reviewing and @ivastar for editing yet another submission! JOSS simply wouldn't be possible without its community of volunteers.
